Extract JVM standard library native functions from duke-interpreter lib.rs to natives.rs
