+++ crates/duke-interpreter/src/lib.rs
@@ -6,6 +6,7 @@
 
 /// Core execution context types for methods and classes.
 pub mod context;
+pub(crate) mod natives;
 /// Repositories for loaded classes and registered native methods.
 pub mod registry;
 mod threading;
@@ -80,14 +81,20 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/io/PrintStream",
         "println",
         "(Ljava/lang/String;)V",
-        native_println_string,
+        natives::native_println_string,
     );
-    registry
-        .natives_mut()
-        .register("java/io/PrintStream", "println", "(I)V", native_println_int);
-    registry
-        .natives_mut()
-        .register("java/io/PrintStream", "println", "()V", native_println_void);
+    registry.natives_mut().register(
+        "java/io/PrintStream",
+        "println",
+        "(I)V",
+        natives::native_println_int,
+    );
+    registry.natives_mut().register(
+        "java/io/PrintStream",
+        "println",
+        "()V",
+        natives::native_println_void,
+    );
 
     let file_ctx = ClassContext {
         class_name: "java/io/File".to_string(),
@@ -109,19 +116,22 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/io/File",
         "<init>",
         "(Ljava/lang/String;)V",
-        native_file_init,
+        natives::native_file_init,
     );
     registry
         .natives_mut()
-        .register("java/io/File", "exists", "()Z", native_file_exists);
-    registry
-        .natives_mut()
-        .register("java/io/File", "isFile", "()Z", native_file_is_file);
+        .register("java/io/File", "exists", "()Z", natives::native_file_exists);
+    registry.natives_mut().register(
+        "java/io/File",
+        "isFile",
+        "()Z",
+        natives::native_file_is_file,
+    );
     registry.natives_mut().register(
         "java/io/File",
         "isDirectory",
         "()Z",
-        native_file_is_directory,
+        natives::native_file_is_directory,
     );
 
     let io_exception_ctx = ClassContext {
@@ -235,25 +245,25 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/io/FileInputStream",
         "<init>",
         "(Ljava/lang/String;)V",
-        native_file_input_stream_init,
+        natives::native_file_input_stream_init,
     );
     registry.natives_mut().register(
         "java/io/FileInputStream",
         "read",
         "()I",
-        native_file_input_stream_read,
+        natives::native_file_input_stream_read,
     );
     registry.natives_mut().register(
         "java/io/FileInputStream",
         "read",
         "([B)I",
-        native_file_input_stream_read_bytes,
+        natives::native_file_input_stream_read_bytes,
     );
     registry.natives_mut().register(
         "java/io/FileInputStream",
         "close",
         "()V",
-        native_file_input_stream_close,
+        natives::native_file_input_stream_close,
     );
 
     let file_output_stream_ctx = ClassContext {
@@ -276,25 +286,25 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/io/FileOutputStream",
         "<init>",
         "(Ljava/lang/String;)V",
-        native_file_output_stream_init,
+        natives::native_file_output_stream_init,
     );
     registry.natives_mut().register(
         "java/io/FileOutputStream",
         "write",
         "(I)V",
-        native_file_output_stream_write,
+        natives::native_file_output_stream_write,
     );
     registry.natives_mut().register(
         "java/io/FileOutputStream",
         "write",
         "([B)V",
-        native_file_output_stream_write_bytes,
+        natives::native_file_output_stream_write_bytes,
     );
     registry.natives_mut().register(
         "java/io/FileOutputStream",
         "close",
         "()V",
-        native_file_output_stream_close,
+        natives::native_file_output_stream_close,
     );
 
     let server_socket_ctx = ClassContext {
@@ -324,25 +334,25 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/net/ServerSocket",
         "<init>",
         "(I)V",
-        native_server_socket_init,
+        natives::native_server_socket_init,
     );
     registry.natives_mut().register(
         "java/net/ServerSocket",
         "accept",
         "()Ljava/net/Socket;",
-        native_server_socket_accept,
+        natives::native_server_socket_accept,
     );
     registry.natives_mut().register(
         "java/net/ServerSocket",
         "getLocalPort",
         "()I",
-        native_server_socket_get_local_port,
+        natives::native_server_socket_get_local_port,
     );
     registry.natives_mut().register(
         "java/net/ServerSocket",
         "close",
         "()V",
-        native_server_socket_close,
+        natives::native_server_socket_close,
     );
 
     let socket_ctx = ClassContext {
@@ -372,23 +382,26 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/net/Socket",
         "<init>",
         "(Ljava/lang/String;I)V",
-        native_socket_init,
+        natives::native_socket_init,
     );
     registry.natives_mut().register(
         "java/net/Socket",
         "getInputStream",
         "()Ljava/io/InputStream;",
-        native_socket_get_input_stream,
+        natives::native_socket_get_input_stream,
     );
     registry.natives_mut().register(
         "java/net/Socket",
         "getOutputStream",
         "()Ljava/io/OutputStream;",
-        native_socket_get_output_stream,
+        natives::native_socket_get_output_stream,
     );
-    registry
-        .natives_mut()
-        .register("java/net/Socket", "close", "()V", native_socket_close);
+    registry.natives_mut().register(
+        "java/net/Socket",
+        "close",
+        "()V",
+        natives::native_socket_close,
+    );
 
     let socket_input_stream_ctx = ClassContext {
         class_name: "duke/net/SocketInputStream".to_string(),
@@ -410,19 +423,19 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "duke/net/SocketInputStream",
         "read",
         "()I",
-        native_file_input_stream_read,
+        natives::native_file_input_stream_read,
     );
     registry.natives_mut().register(
         "duke/net/SocketInputStream",
         "read",
         "([B)I",
-        native_file_input_stream_read_bytes,
+        natives::native_file_input_stream_read_bytes,
     );
     registry.natives_mut().register(
         "duke/net/SocketInputStream",
         "close",
         "()V",
-        native_file_input_stream_close,
+        natives::native_file_input_stream_close,
     );
 
     let socket_output_stream_ctx = ClassContext {
@@ -445,19 +458,19 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "duke/net/SocketOutputStream",
         "write",
         "(I)V",
-        native_file_output_stream_write,
+        natives::native_file_output_stream_write,
     );
     registry.natives_mut().register(
         "duke/net/SocketOutputStream",
         "write",
         "([B)V",
-        native_file_output_stream_write_bytes,
+        natives::native_file_output_stream_write_bytes,
     );
     registry.natives_mut().register(
         "duke/net/SocketOutputStream",
         "close",
         "()V",
-        native_file_output_stream_close,
+        natives::native_file_output_stream_close,
     );
 
     // Register java/lang/String ClassContext (empty — instance methods are native).
@@ -479,122 +492,131 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
     registry.register(string_ctx);
 
     // String instance methods
-    registry
-        .natives_mut()
-        .register("java/lang/String", "length", "()I", native_string_length);
+    registry.natives_mut().register(
+        "java/lang/String",
+        "length",
+        "()I",
+        natives::native_string_length,
+    );
     registry.natives_mut().register(
         "java/lang/String",
         "equals",
         "(Ljava/lang/Object;)Z",
-        native_string_equals,
+        natives::native_string_equals,
     );
-    registry
-        .natives_mut()
-        .register("java/lang/String", "charAt", "(I)C", native_string_char_at);
+    registry.natives_mut().register(
+        "java/lang/String",
+        "charAt",
+        "(I)C",
+        natives::native_string_char_at,
+    );
     registry.natives_mut().register(
         "java/lang/String",
         "substring",
         "(I)Ljava/lang/String;",
-        native_string_substring,
+        natives::native_string_substring,
     );
     registry.natives_mut().register(
         "java/lang/String",
         "substring",
         "(II)Ljava/lang/String;",
-        native_string_substring_range,
+        natives::native_string_substring_range,
     );
     registry.natives_mut().register(
         "java/lang/String",
         "indexOf",
         "(Ljava/lang/String;)I",
-        native_string_indexof,
+        natives::native_string_indexof,
     );
     registry.natives_mut().register(
         "java/lang/String",
         "contains",
         "(Ljava/lang/CharSequence;)Z",
-        native_string_contains,
+        natives::native_string_contains,
     );
-    registry
-        .natives_mut()
-        .register("java/lang/String", "isEmpty", "()Z", native_string_isempty);
+    registry.natives_mut().register(
+        "java/lang/String",
+        "isEmpty",
+        "()Z",
+        natives::native_string_isempty,
+    );
     registry.natives_mut().register(
         "java/lang/String",
         "compareTo",
         "(Ljava/lang/String;)I",
-        native_string_compareto,
+        natives::native_string_compareto,
     );
     registry.natives_mut().register(
         "java/lang/String",
         "compareTo",
         "(Ljava/lang/Object;)I",
-        native_string_compareto_object,
+        natives::native_string_compareto_object,
     );
     registry.natives_mut().register(
         "java/lang/String",
         "startsWith",
         "(Ljava/lang/String;)Z",
-        native_string_startswith,
+        natives::native_string_startswith,
     );
     registry.natives_mut().register(
         "java/lang/String",
         "endsWith",
         "(Ljava/lang/String;)Z",
-        native_string_endswith,
+        natives::native_string_endswith,
     );
     registry.natives_mut().register(
         "java/lang/String",
         "trim",
         "()Ljava/lang/String;",
-        native_string_trim,
+        natives::native_string_trim,
     );
     registry.natives_mut().register(
         "java/lang/String",
         "toCharArray",
         "()[C",
-        native_string_tochararray,
+        natives::native_string_tochararray,
     );
     registry.natives_mut().register(
         "java/lang/String",
         "toUpperCase",
         "()Ljava/lang/String;",
-        native_string_touppercase,
+        natives::native_string_touppercase,
     );
     registry.natives_mut().register(
         "java/lang/String",
         "toLowerCase",
         "()Ljava/lang/String;",
-        native_string_tolowercase,
+        natives::native_string_tolowercase,
     );
     registry.natives_mut().register(
         "java/lang/String",
         "replace",
         "(CC)Ljava/lang/String;",
-        native_string_replace_char,
+        natives::native_string_replace_char,
     );
     registry.natives_mut().register(
         "java/lang/String",
         "replace",
         "(Ljava/lang/CharSequence;Ljava/lang/CharSequence;)Ljava/lang/String;",
-        native_string_replace_charsequence,
+        natives::native_string_replace_charsequence,
     );
     registry.natives_mut().register(
         "java/lang/String",
         "split",
         "(Ljava/lang/String;)[Ljava/lang/String;",
-        native_string_split,
+        natives::native_string_split,
     );
     registry.natives_mut().register(
         "java/lang/String",
         "hashCode",
         "()I",
-        native_string_hashcode,
+        natives::native_string_hashcode,
     );
     registry.natives_mut().register(
         "java/lang/String",
         "toString",
         "()Ljava/lang/String;",
-        native_string_tostring,
+        natives::native_string_tostring,
     );
 
     // String static methods
@@ -602,7 +624,7 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/lang/String",
         "valueOf",
         "(I)Ljava/lang/String;",
-        native_string_value_of_int,
+        natives::native_string_value_of_int,
     );
 
     // PrintStream.print (no newline)
@@ -610,77 +632,98 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/io/PrintStream",
         "print",
         "(Ljava/lang/String;)V",
-        native_print_string,
+        natives::native_print_string,
     );
-    registry
-        .natives_mut()
-        .register("java/io/PrintStream", "print", "(I)V", native_print_int);
+    registry.natives_mut().register(
+        "java/io/PrintStream",
+        "print",
+        "(I)V",
+        natives::native_print_int,
+    );
 
     // println overloads
     registry.natives_mut().register(
         "java/io/PrintStream",
         "println",
         "(J)V",
-        native_println_long,
+        natives::native_println_long,
     );
     registry.natives_mut().register(
         "java/io/PrintStream",
         "println",
         "(F)V",
-        native_println_float,
+        natives::native_println_float,
     );
     registry.natives_mut().register(
         "java/io/PrintStream",
         "println",
         "(D)V",
-        native_println_double,
+        natives::native_println_double,
     );
     registry.natives_mut().register(
         "java/io/PrintStream",
         "println",
         "(Z)V",
-        native_println_boolean,
+        natives::native_println_boolean,
     );
     registry.natives_mut().register(
         "java/io/PrintStream",
         "println",
         "(C)V",
-        native_println_char,
+        natives::native_println_char,
     );
     registry.natives_mut().register(
         "java/io/PrintStream",
         "println",
         "(Ljava/lang/Object;)V",
-        native_println_object,
+        natives::native_println_object,
     );
 
     // print overloads
-    registry
-        .natives_mut()
-        .register("java/io/PrintStream", "print", "(J)V", native_print_long);
-    registry
-        .natives_mut()
-        .register("java/io/PrintStream", "print", "(F)V", native_print_float);
-    registry
-        .natives_mut()
-        .register("java/io/PrintStream", "print", "(D)V", native_print_double);
-    registry
-        .natives_mut()
-        .register("java/io/PrintStream", "print", "(Z)V", native_print_boolean);
-    registry
-        .natives_mut()
-        .register("java/io/PrintStream", "print", "(C)V", native_print_char);
+    registry.natives_mut().register(
+        "java/io/PrintStream",
+        "print",
+        "(J)V",
+        natives::native_print_long,
+    );
+    registry.natives_mut().register(
+        "java/io/PrintStream",
+        "print",
+        "(F)V",
+        natives::native_print_float,
+    );
+    registry.natives_mut().register(
+        "java/io/PrintStream",
+        "print",
+        "(D)V",
+        natives::native_print_double,
+    );
+    registry.natives_mut().register(
+        "java/io/PrintStream",
+        "print",
+        "(Z)V",
+        natives::native_print_boolean,
+    );
+    registry.natives_mut().register(
+        "java/io/PrintStream",
+        "print",
+        "(C)V",
+        natives::native_print_char,
+    );
     registry.natives_mut().register(
         "java/io/PrintStream",
         "print",
         "(Ljava/lang/Object;)V",
-        native_print_object,
+        natives::native_print_object,
     );
 
     // System.exit (static)
-    registry
-        .natives_mut()
-        .register("java/lang/System", "exit", "(I)V", native_system_exit);
+    registry.natives_mut().register(
+        "java/lang/System",
+        "exit",
+        "(I)V",
+        natives::native_system_exit,
+    );
 
     // Register synthetic exception hierarchy so is_assignable_from can walk it.
     // java/lang/Object (root — no super)
@@ -702,19 +745,19 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/lang/Object",
         "hashCode",
         "()I",
-        native_object_hashcode,
+        natives::native_object_hashcode,
     );
     registry.natives_mut().register(
         "java/lang/Object",
         "toString",
         "()Ljava/lang/String;",
-        native_object_tostring,
+        natives::native_object_tostring,
     );
     registry.natives_mut().register(
         "java/lang/Object",
         "clone",
         "()Ljava/lang/Object;",
-        native_object_clone,
+        natives::native_object_clone,
     );
 
     // java/lang/Class — lightweight stub for class literals
@@ -767,24 +810,36 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         bootstrap_methods: Vec::new(),
     };
     registry.register(thread_ctx);
-    registry
-        .natives_mut()
-        .register("java/lang/Thread", "<init>", "()V", native_thread_init);
+    registry.natives_mut().register(
+        "java/lang/Thread",
+        "<init>",
+        "()V",
+        natives::native_thread_init,
+    );
     registry.natives_mut().register(
         "java/lang/Thread",
         "<init>",
         "(Ljava/lang/Runnable;)V",
-        native_thread_init_runnable,
+        natives::native_thread_init_runnable,
     );
-    registry
-        .natives_mut()
-        .register("java/lang/Thread", "start", "()V", native_thread_start);
-    registry
-        .natives_mut()
-        .register("java/lang/Thread", "join", "()V", native_thread_join);
-    registry
-        .natives_mut()
-        .register("java/lang/Thread", "sleep", "(J)V", native_thread_sleep);
+    registry.natives_mut().register(
+        "java/lang/Thread",
+        "start",
+        "()V",
+        natives::native_thread_start,
+    );
+    registry.natives_mut().register(
+        "java/lang/Thread",
+        "join",
+        "()V",
+        natives::native_thread_join,
+    );
+    registry.natives_mut().register(
+        "java/lang/Thread",
+        "sleep",
+        "(J)V",
+        natives::native_thread_sleep,
+    );
 
     // java/lang/Throwable extends Object
     let throwable_ctx = ClassContext {
@@ -803,7 +858,7 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/lang/Throwable",
         "addSuppressed",
         "(Ljava/lang/Throwable;)V",
-        native_throwable_add_suppressed,
+        natives::native_throwable_add_suppressed,
     );
 
     // java/lang/Exception extends Throwable
@@ -880,22 +935,25 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/lang/Enum",
         "<init>",
         "(Ljava/lang/String;I)V",
-        native_enum_init,
+        natives::native_enum_init,
     );
-    registry
-        .natives_mut()
-        .register("java/lang/Enum", "ordinal", "()I", native_enum_ordinal);
+    registry.natives_mut().register(
+        "java/lang/Enum",
+        "ordinal",
+        "()I",
+        natives::native_enum_ordinal,
+    );
     registry.natives_mut().register(
         "java/lang/Enum",
         "name",
         "()Ljava/lang/String;",
-        native_enum_name,
+        natives::native_enum_name,
     );
     registry.natives_mut().register(
         "java/lang/Enum",
         "valueOf",
         "(Ljava/lang/Class;Ljava/lang/String;)Ljava/lang/Enum;",
-        native_enum_valueof,
+        natives::native_enum_valueof,
     );
 
     // java/lang/Integer — boxed int with value field + numeric constants
@@ -931,31 +989,31 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/lang/Integer",
         "parseInt",
         "(Ljava/lang/String;)I",
-        native_integer_parseint,
+        natives::native_integer_parseint,
     );
     registry.natives_mut().register(
         "java/lang/Integer",
         "valueOf",
         "(I)Ljava/lang/Integer;",
-        native_integer_valueof,
+        natives::native_integer_valueof,
     );
     registry.natives_mut().register(
         "java/lang/Integer",
         "intValue",
         "()I",
-        native_integer_intvalue,
+        natives::native_integer_intvalue,
     );
     registry.natives_mut().register(
         "java/lang/Integer",
         "toString",
         "(I)Ljava/lang/String;",
-        native_integer_tostring_static,
+        natives::native_integer_tostring_static,
     );
     registry.natives_mut().register(
         "java/lang/Integer",
         "compareTo",
         "(Ljava/lang/Object;)I",
-        native_integer_compareto,
+        natives::native_integer_compareto,
     );
 
     // java/lang/Long — boxed long with value field + numeric constants
@@ -991,28 +1049,31 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/lang/Long",
         "parseLong",
         "(Ljava/lang/String;)J",
-        native_long_parselong,
+        natives::native_long_parselong,
     );
     registry.natives_mut().register(
         "java/lang/Long",
         "valueOf",
         "(J)Ljava/lang/Long;",
-        native_long_valueof,
+        natives::native_long_valueof,
     );
-    registry
-        .natives_mut()
-        .register("java/lang/Long", "longValue", "()J", native_long_longvalue);
+    registry.natives_mut().register(
+        "java/lang/Long",
+        "longValue",
+        "()J",
+        natives::native_long_longvalue,
+    );
     registry.natives_mut().register(
         "java/lang/Long",
         "toString",
         "(J)Ljava/lang/String;",
-        native_long_tostring_static,
+        natives::native_long_tostring_static,
     );
     registry.natives_mut().register(
         "java/lang/Long",
         "compareTo",
         "(Ljava/lang/Object;)I",
-        native_long_compareto,
+        natives::native_long_compareto,
     );
 
     // java/lang/Double — boxed double with value field + numeric constants
@@ -1069,28 +1130,31 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/lang/Double",
         "parseDouble",
         "(Ljava/lang/String;)D",
-        native_double_parsedouble,
+        natives::native_double_parsedouble,
     );
     registry.natives_mut().register(
         "java/lang/Double",
         "valueOf",
         "(D)Ljava/lang/Double;",
-        native_double_valueof,
+        natives::native_double_valueof,
     );
     registry.natives_mut().register(
         "java/lang/Double",
         "doubleValue",
         "()D",
-        native_double_doublevalue,
+        natives::native_double_doublevalue,
     );
-    registry
-        .natives_mut()
-        .register("java/lang/Double", "isNaN", "(D)Z", native_double_isnan);
+    registry.natives_mut().register(
+        "java/lang/Double",
+        "isNaN",
+        "(D)Z",
+        natives::native_double_isnan,
+    );
     registry.natives_mut().register(
         "java/lang/Double",
         "compareTo",
         "(Ljava/lang/Object;)I",
-        native_double_compareto,
+        natives::native_double_compareto,
     );
 
     // java/lang/Float — boxed float with value field
@@ -1114,7 +1178,7 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/lang/Float",
         "parseFloat",
         "(Ljava/lang/String;)F",
-        native_float_parsefloat,
+        natives::native_float_parsefloat,
     );
 
     // java/lang/Boolean — static utility
@@ -1134,7 +1198,7 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/lang/Boolean",
         "parseBoolean",
         "(Ljava/lang/String;)Z",
-        native_boolean_parseboolean,
+        natives::native_boolean_parseboolean,
     );
 
     // java/lang/Math — static math utilities with PI and E constants
@@ -1164,85 +1228,118 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         bootstrap_methods: Vec::new(),
     };
     registry.register(math_ctx);
+    registry.natives_mut().register(
+        "java/lang/Math",
+        "max",
+        "(II)I",
+        natives::native_math_max_int,
+    );
+    registry.natives_mut().register(
+        "java/lang/Math",
+        "min",
+        "(II)I",
+        natives::native_math_min_int,
+    );
+    registry.natives_mut().register(
+        "java/lang/Math",
+        "abs",
+        "(I)I",
+        natives::native_math_abs_int,
+    );
     registry
         .natives_mut()
-        .register("java/lang/Math", "max", "(II)I", native_math_max_int);
+        .register("java/lang/Math", "sqrt", "(D)D", natives::native_math_sqrt);
     registry
         .natives_mut()
-        .register("java/lang/Math", "min", "(II)I", native_math_min_int);
+        .register("java/lang/Math", "pow", "(DD)D", natives::native_math_pow);
+    registry.natives_mut().register(
+        "java/lang/Math",
+        "floor",
+        "(D)D",
+        natives::native_math_floor,
+    );
     registry
         .natives_mut()
-        .register("java/lang/Math", "abs", "(I)I", native_math_abs_int);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "sqrt", "(D)D", native_math_sqrt);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "pow", "(DD)D", native_math_pow);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "floor", "(D)D", native_math_floor);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "ceil", "(D)D", native_math_ceil);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "round", "(D)J", native_math_round_double);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "abs", "(J)J", native_math_abs_long);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "abs", "(D)D", native_math_abs_double);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "max", "(JJ)J", native_math_max_long);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "min", "(JJ)J", native_math_min_long);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "max", "(DD)D", native_math_max_double);
-    registry
-        .natives_mut()
-        .register("java/lang/Math", "min", "(DD)D", native_math_min_double);
+        .register("java/lang/Math", "ceil", "(D)D", natives::native_math_ceil);
+    registry.natives_mut().register(
+        "java/lang/Math",
+        "round",
+        "(D)J",
+        natives::native_math_round_double,
+    );
+    registry.natives_mut().register(
+        "java/lang/Math",
+        "abs",
+        "(J)J",
+        natives::native_math_abs_long,
+    );
+    registry.natives_mut().register(
+        "java/lang/Math",
+        "abs",
+        "(D)D",
+        natives::native_math_abs_double,
+    );
+    registry.natives_mut().register(
+        "java/lang/Math",
+        "max",
+        "(JJ)J",
+        natives::native_math_max_long,
+    );
+    registry.natives_mut().register(
+        "java/lang/Math",
+        "min",
+        "(JJ)J",
+        natives::native_math_min_long,
+    );
+    registry.natives_mut().register(
+        "java/lang/Math",
+        "max",
+        "(DD)D",
+        natives::native_math_max_double,
+    );
+    registry.natives_mut().register(
+        "java/lang/Math",
+        "min",
+        "(DD)D",
+        natives::native_math_min_double,
+    );
 
     // String.valueOf overloads (int already registered above)
     registry.natives_mut().register(
         "java/lang/String",
         "valueOf",
         "(J)Ljava/lang/String;",
-        native_string_value_of_long,
+        natives::native_string_value_of_long,
     );
     registry.natives_mut().register(
         "java/lang/String",
         "valueOf",
         "(D)Ljava/lang/String;",
-        native_string_value_of_double,
+        natives::native_string_value_of_double,
     );
     registry.natives_mut().register(
         "java/lang/String",
         "valueOf",
         "(F)Ljava/lang/String;",
-        native_string_value_of_float,
+        natives::native_string_value_of_float,
     );
     registry.natives_mut().register(
         "java/lang/String",
         "valueOf",
         "(Z)Ljava/lang/String;",
-        native_string_value_of_boolean,
+        natives::native_string_value_of_boolean,
     );
     registry.natives_mut().register(
         "java/lang/String",
         "valueOf",
         "(C)Ljava/lang/String;",
-        native_string_value_of_char,
+        natives::native_string_value_of_char,
     );
     registry.natives_mut().register(
         "java/lang/String",
         "valueOf",
         "(Ljava/lang/Object;)Ljava/lang/String;",
-        native_string_value_of_object,
+        natives::native_string_value_of_object,
     );
 
     // String.concat
@@ -1250,7 +1347,7 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/lang/String",
         "concat",
         "(Ljava/lang/String;)Ljava/lang/String;",
-        native_string_concat,
+        natives::native_string_concat,
     );
 
     // String.format(String, Object[]) -> String
@@ -1258,7 +1355,7 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/lang/String",
         "format",
         "(Ljava/lang/String;[Ljava/lang/Object;)Ljava/lang/String;",
-        native_string_format,
+        natives::native_string_format,
     );
 
     // java/lang/StringBuilder — mutable string buffer
@@ -1276,76 +1373,82 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
     registry.register(sb_ctx);
 
     // StringBuilder.<init>()V
-    registry
-        .natives_mut()
-        .register("java/lang/StringBuilder", "<init>", "()V", native_sb_init);
+    registry.natives_mut().register(
+        "java/lang/StringBuilder",
+        "<init>",
+        "()V",
+        natives::native_sb_init,
+    );
     // StringBuilder.<init>(String)V
     registry.natives_mut().register(
         "java/lang/StringBuilder",
         "<init>",
         "(Ljava/lang/String;)V",
-        native_sb_init_string,
+        natives::native_sb_init_string,
     );
     // StringBuilder.append(String)
     registry.natives_mut().register(
         "java/lang/StringBuilder",
         "append",
         "(Ljava/lang/String;)Ljava/lang/StringBuilder;",
-        native_sb_append_string,
+        natives::native_sb_append_string,
     );
     // StringBuilder.append(int)
     registry.natives_mut().register(
         "java/lang/StringBuilder",
         "append",
         "(I)Ljava/lang/StringBuilder;",
-        native_sb_append_int,
+        natives::native_sb_append_int,
     );
     // StringBuilder.append(long)
     registry.natives_mut().register(
         "java/lang/StringBuilder",
         "append",
         "(J)Ljava/lang/StringBuilder;",
-        native_sb_append_long,
+        natives::native_sb_append_long,
     );
     // StringBuilder.append(double)
     registry.natives_mut().register(
         "java/lang/StringBuilder",
         "append",
         "(D)Ljava/lang/StringBuilder;",
-        native_sb_append_double,
+        natives::native_sb_append_double,
     );
     // StringBuilder.append(float)
     registry.natives_mut().register(
         "java/lang/StringBuilder",
         "append",
         "(F)Ljava/lang/StringBuilder;",
-        native_sb_append_float,
+        natives::native_sb_append_float,
     );
     // StringBuilder.append(boolean)
     registry.natives_mut().register(
         "java/lang/StringBuilder",
         "append",
         "(Z)Ljava/lang/StringBuilder;",
-        native_sb_append_boolean,
+        natives::native_sb_append_boolean,
     );
     // StringBuilder.append(char)
     registry.natives_mut().register(
         "java/lang/StringBuilder",
         "append",
         "(C)Ljava/lang/StringBuilder;",
-        native_sb_append_char,
+        natives::native_sb_append_char,
     );
     // StringBuilder.toString()
     registry.natives_mut().register(
         "java/lang/StringBuilder",
         "toString",
         "()Ljava/lang/String;",
-        native_sb_tostring,
+        natives::native_sb_tostring,
     );
     // StringBuilder.length()
-    registry
-        .natives_mut()
-        .register("java/lang/StringBuilder", "length", "()I", native_sb_length);
+    registry.natives_mut().register(
+        "java/lang/StringBuilder",
+        "length",
+        "()I",
+        natives::native_sb_length,
+    );
 
     // java/lang/Character — static character utilities + boxed char
     let character_ctx = ClassContext {
@@ -1370,70 +1473,70 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/lang/Character",
         "isDigit",
         "(C)Z",
-        native_char_is_digit,
+        natives::native_char_is_digit,
     );
     // Character.isLetter(C)Z
     registry.natives_mut().register(
         "java/lang/Character",
         "isLetter",
         "(C)Z",
-        native_char_is_letter,
+        natives::native_char_is_letter,
     );
     // Character.isWhitespace(C)Z
     registry.natives_mut().register(
         "java/lang/Character",
         "isWhitespace",
         "(C)Z",
-        native_char_is_whitespace,
+        natives::native_char_is_whitespace,
     );
     // Character.isUpperCase(C)Z
     registry.natives_mut().register(
         "java/lang/Character",
         "isUpperCase",
         "(C)Z",
-        native_char_is_uppercase,
+        natives::native_char_is_uppercase,
     );
     // Character.isLowerCase(C)Z
     registry.natives_mut().register(
         "java/lang/Character",
         "isLowerCase",
         "(C)Z",
-        native_char_is_lowercase,
+        natives::native_char_is_lowercase,
     );
     // Character.toUpperCase(C)C
     registry.natives_mut().register(
         "java/lang/Character",
         "toUpperCase",
         "(C)C",
-        native_char_to_uppercase,
+        natives::native_char_to_uppercase,
     );
     // Character.toLowerCase(C)C
     registry.natives_mut().register(
         "java/lang/Character",
         "toLowerCase",
         "(C)C",
-        native_char_to_lowercase,
+        natives::native_char_to_lowercase,
     );
     // Character.isLetterOrDigit(C)Z
     registry.natives_mut().register(
         "java/lang/Character",
         "isLetterOrDigit",
         "(C)Z",
-        native_char_is_letter_or_digit,
+        natives::native_char_is_letter_or_digit,
     );
     // Character.valueOf(C)Ljava/lang/Character;
     registry.natives_mut().register(
         "java/lang/Character",
         "valueOf",
         "(C)Ljava/lang/Character;",
-        native_char_valueof,
+        natives::native_char_valueof,
     );
     // Character.charValue()C
     registry.natives_mut().register(
         "java/lang/Character",
         "charValue",
         "()C",
-        native_char_charvalue,
+        natives::native_char_charvalue,
     );
 
     // java/util/ArrayList — dynamic list backed by growable fields
@@ -1462,34 +1565,37 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/util/ArrayList",
         "<init>",
         "()V",
-        native_arraylist_init,
+        natives::native_arraylist_init,
     );
     registry.natives_mut().register(
         "java/util/ArrayList",
         "add",
         "(Ljava/lang/Object;)Z",
-        native_arraylist_add,
+        natives::native_arraylist_add,
     );
     registry.natives_mut().register(
         "java/util/ArrayList",
         "get",
         "(I)Ljava/lang/Object;",
-        native_arraylist_get,
+        natives::native_arraylist_get,
     );
-    registry
-        .natives_mut()
-        .register("java/util/ArrayList", "size", "()I", native_arraylist_size);
+    registry.natives_mut().register(
+        "java/util/ArrayList",
+        "size",
+        "()I",
+        natives::native_arraylist_size,
+    );
     registry.natives_mut().register(
         "java/util/ArrayList",
         "iterator",
         "()Ljava/util/Iterator;",
-        native_arraylist_iterator,
+        natives::native_arraylist_iterator,
     );
     registry.natives_mut().register_callback(
         "java/util/ArrayList",
         "sort",
         "(Ljava/util/Comparator;)V",
-        array_list_sort,
+        natives::array_list_sort,
     );
 
     // duke/util/ArrayListIterator — internal iterator for ArrayList
@@ -1521,19 +1627,19 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "duke/util/ArrayListIterator",
         "<init>",
         "()V",
-        native_arraylist_iter_init,
+        natives::native_arraylist_iter_init,
     );
     registry.natives_mut().register(
         "duke/util/ArrayListIterator",
         "hasNext",
         "()Z",
-        native_arraylist_iter_hasnext,
+        natives::native_arraylist_iter_hasnext,
     );
     registry.natives_mut().register(
         "duke/util/ArrayListIterator",
         "next",
         "()Ljava/lang/Object;",
-        native_arraylist_iter_next,
+        natives::native_arraylist_iter_next,
     );
 
     // java/util/Arrays — static array utilities
@@ -1549,30 +1655,36 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         bootstrap_methods: Vec::new(),
     };
     registry.register(arrays_ctx);
-    registry
-        .natives_mut()
-        .register("java/util/Arrays", "fill", "([II)V", native_arrays_fill_int);
+    registry.natives_mut().register(
+        "java/util/Arrays",
+        "fill",
+        "([II)V",
+        natives::native_arrays_fill_int,
+    );
     registry.natives_mut().register(
         "java/util/Arrays",
         "fill",
         "([Ljava/lang/Object;Ljava/lang/Object;)V",
-        native_arrays_fill_object,
+        natives::native_arrays_fill_object,
     );
     registry.natives_mut().register(
         "java/util/Arrays",
         "copyOf",
         "([II)[I",
-        native_arrays_copyof_int,
+        natives::native_arrays_copyof_int,
     );
     registry.natives_mut().register(
         "java/util/Arrays",
         "copyOf",
         "([Ljava/lang/Object;I)[Ljava/lang/Object;",
-        native_arrays_copyof_object,
+        natives::native_arrays_copyof_object,
     );
-    registry
-        .natives_mut()
-        .register("java/util/Arrays", "sort", "([I)V", native_arrays_sort_int);
+    registry.natives_mut().register(
+        "java/util/Arrays",
+        "sort",
+        "([I)V",
+        natives::native_arrays_sort_int,
+    );
 
     // java/util/HashMap — hash map backed by flat key/value pair list in fields
     // fields[0] = Int(size), fields[1]=key0, fields[2]=val0, fields[3]=key1, ...
@@ -1592,47 +1704,53 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         bootstrap_methods: Vec::new(),
     };
     registry.register(hashmap_ctx);
-    registry
-        .natives_mut()
-        .register("java/util/HashMap", "<init>", "()V", native_hashmap_init);
+    registry.natives_mut().register(
+        "java/util/HashMap",
+        "<init>",
+        "()V",
+        natives::native_hashmap_init,
+    );
     registry.natives_mut().register(
         "java/util/HashMap",
         "put",
         "(Ljava/lang/Object;Ljava/lang/Object;)Ljava/lang/Object;",
-        native_hashmap_put,
+        natives::native_hashmap_put,
     );
     registry.natives_mut().register(
         "java/util/HashMap",
         "get",
         "(Ljava/lang/Object;)Ljava/lang/Object;",
-        native_hashmap_get,
+        natives::native_hashmap_get,
     );
     registry.natives_mut().register(
         "java/util/HashMap",
         "containsKey",
         "(Ljava/lang/Object;)Z",
-        native_hashmap_contains_key,
+        natives::native_hashmap_contains_key,
     );
-    registry
-        .natives_mut()
-        .register("java/util/HashMap", "size", "()I", native_hashmap_size);
+    registry.natives_mut().register(
+        "java/util/HashMap",
+        "size",
+        "()I",
+        natives::native_hashmap_size,
+    );
     registry.natives_mut().register(
         "java/util/HashMap",
         "remove",
         "(Ljava/lang/Object;)Ljava/lang/Object;",
-        native_hashmap_remove,
+        natives::native_hashmap_remove,
     );
     registry.natives_mut().register(
         "java/util/HashMap",
         "isEmpty",
         "()Z",
-        native_hashmap_is_empty,
+        natives::native_hashmap_is_empty,
     );
     registry.natives_mut().register(
         "java/util/HashMap",
         "getOrDefault",
         "(Ljava/lang/Object;Ljava/lang/Object;)Ljava/lang/Object;",
-        native_hashmap_get_or_default,
+        natives::native_hashmap_get_or_default,
     );
 
     // java/util/HashSet — set backed by unique elements in fields
@@ -1653,35 +1771,41 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         bootstrap_methods: Vec::new(),
     };
     registry.register(hashset_ctx);
-    registry
-        .natives_mut()
-        .register("java/util/HashSet", "<init>", "()V", native_hashset_init);
+    registry.natives_mut().register(
+        "java/util/HashSet",
+        "<init>",
+        "()V",
+        natives::native_hashset_init,
+    );
     registry.natives_mut().register(
         "java/util/HashSet",
         "add",
         "(Ljava/lang/Object;)Z",
-        native_hashset_add,
+        natives::native_hashset_add,
     );
     registry.natives_mut().register(
         "java/util/HashSet",
         "contains",
         "(Ljava/lang/Object;)Z",
-        native_hashset_contains,
+        natives::native_hashset_contains,
     );
     registry.natives_mut().register(
         "java/util/HashSet",
         "remove",
         "(Ljava/lang/Object;)Z",
-        native_hashset_remove,
+        natives::native_hashset_remove,
     );
-    registry
-        .natives_mut()
-        .register("java/util/HashSet", "size", "()I", native_hashset_size);
+    registry.natives_mut().register(
+        "java/util/HashSet",
+        "size",
+        "()I",
+        natives::native_hashset_size,
+    );
     registry.natives_mut().register(
         "java/util/HashSet",
         "isEmpty",
         "()Z",
-        native_hashset_is_empty,
+        natives::native_hashset_is_empty,
     );
 
     // java/util/Collections — static utility class
@@ -1702,7 +1826,7 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/util/Collections",
         "sort",
         "(Ljava/util/List;)V",
-        native_collections_sort,
+        natives::native_collections_sort,
     );
 
     // ── java.util.zip ──────────────────────────────────────────────────
@@ -1769,25 +1893,25 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/util/zip/ZipEntry",
         "getName",
         "()Ljava/lang/String;",
-        native_zip_entry_get_name,
+        natives::native_zip_entry_get_name,
     );
     registry.natives_mut().register(
         "java/util/zip/ZipEntry",
         "getCompressedSize",
         "()J",
-        native_zip_entry_get_compressed_size,
+        natives::native_zip_entry_get_compressed_size,
     );
     registry.natives_mut().register(
         "java/util/zip/ZipEntry",
         "getSize",
         "()J",
-        native_zip_entry_get_size,
+        natives::native_zip_entry_get_size,
     );
     registry.natives_mut().register(
         "java/util/zip/ZipEntry",
         "getMethod",
         "()I",
-        native_zip_entry_get_method,
+        natives::native_zip_entry_get_method,
     );
 
     // ZipFile: [fd: Int] → 1 field
@@ -1811,29 +1935,32 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "java/util/zip/ZipFile",
         "<init>",
         "(Ljava/lang/String;)V",
-        native_zip_file_init,
+        natives::native_zip_file_init,
     );
     registry.natives_mut().register(
         "java/util/zip/ZipFile",
         "getEntry",
         "(Ljava/lang/String;)Ljava/util/zip/ZipEntry;",
-        native_zip_file_get_entry,
+        natives::native_zip_file_get_entry,
     );
     registry.natives_mut().register(
         "java/util/zip/ZipFile",
         "getInputStream",
         "(Ljava/util/zip/ZipEntry;)Ljava/io/InputStream;",
-        native_zip_file_get_input_stream,
+        natives::native_zip_file_get_input_stream,
     );
     registry.natives_mut().register(
         "java/util/zip/ZipFile",
         "close",
         "()V",
-        native_zip_file_close,
+        natives::native_zip_file_close,
     );
-    registry
-        .natives_mut()
-        .register("java/util/zip/ZipFile", "size", "()I", native_zip_file_size);
+    registry.natives_mut().register(
+        "java/util/zip/ZipFile",
+        "size",
+        "()I",
+        natives::native_zip_file_size,
+    );
 
     // JarFile extends ZipFile — inherits everything for now.
     let jar_file_ctx = ClassContext {
@@ -1870,3074 +1997,20 @@ pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) 
         "duke/zip/ByteBufferInputStream",
         "read",
         "()I",
-        native_file_input_stream_read,
+        natives::native_file_input_stream_read,
     );
     registry.natives_mut().register(
         "duke/zip/ByteBufferInputStream",
         "read",
         "([B)I",
-        native_file_input_stream_read_bytes,
+        natives::native_file_input_stream_read_bytes,
     );
     registry.natives_mut().register(
         "duke/zip/ByteBufferInputStream",
         "close",
         "()V",
-        native_file_input_stream_close,
+        natives::native_file_input_stream_close,
     );
-}
-
-fn native_println_string(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let string_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => {
-            writeln!(out, "null").ok();
-            return Ok(None);
-        }
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    let obj = heap.get(string_ref)?;
-    let text = obj.string_value.as_deref().unwrap_or("null");
-    writeln!(out, "{text}").ok();
-    Ok(None)
-}
-
-fn native_println_int(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    writeln!(out, "{val}").ok();
-    Ok(None)
-}
-
-#[allow(clippy::unnecessary_wraps)] // must match NativeHandler signature
-fn native_println_void(
-    _args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    writeln!(out).ok();
-    Ok(None)
-}
-
-fn path_from_string_slot(
-    args: &[Slot],
-    idx: usize,
-    heap: &duke_gc::Heap,
-) -> VmResult<std::path::PathBuf> {
-    let path_ref = match args.get(idx) {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let path = heap
-        .get(path_ref)?
-        .string_value
-        .clone()
-        .ok_or(VmError::NullPointerException)?;
-    Ok(std::path::PathBuf::from(path))
-}
-
-fn file_path_from_this(args: &[Slot], heap: &duke_gc::Heap) -> VmResult<std::path::PathBuf> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let path_ref = match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let path = heap
-        .get(path_ref)?
-        .string_value
-        .clone()
-        .ok_or(VmError::NullPointerException)?;
-    Ok(std::path::PathBuf::from(path))
-}
-
-fn native_file_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let path_slot = args.get(1).copied().unwrap_or(Slot::Reference(None));
-    let file_obj = heap.get_mut(this_ref)?;
-    let Some(path_field) = file_obj.fields.first_mut() else {
-        return Err(VmError::InvalidRef { address: this_ref });
-    };
-    *path_field = path_slot;
-    Ok(None)
-}
-
-fn native_file_exists(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let path = file_path_from_this(args, heap)?;
-    Ok(Some(Slot::Int(i32::from(path.exists()))))
-}
-
-fn native_file_is_file(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let path = file_path_from_this(args, heap)?;
-    Ok(Some(Slot::Int(i32::from(path.is_file()))))
-}
-
-fn native_file_is_directory(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let path = file_path_from_this(args, heap)?;
-    Ok(Some(Slot::Int(i32::from(path.is_dir()))))
-}
-
-fn file_stream_id_from_this(args: &[Slot], heap: &duke_gc::Heap) -> VmResult<i32> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(id)) if *id > 0 => Ok(*id),
-        _ => Err(VmError::JavaException {
-            class_name: "java/io/IOException".to_string(),
-        }),
-    }
-}
-
-fn native_file_input_stream_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let path = path_from_string_slot(args, 1, heap)?;
-    let file_id = heap.open_host_input_file(&path)?;
-    let stream_obj = heap.get_mut(this_ref)?;
-    let Some(fd_field) = stream_obj.fields.first_mut() else {
-        return Err(VmError::InvalidRef { address: this_ref });
-    };
-    *fd_field = Slot::Int(file_id);
-    Ok(None)
-}
-
-fn native_file_input_stream_read(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let file_id = file_stream_id_from_this(args, heap)?;
-    Ok(Some(Slot::Int(heap.read_host_file_byte(file_id)?)))
-}
-
-fn native_file_input_stream_read_bytes(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let file_id = file_stream_id_from_this(args, heap)?;
-    let array_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let len = heap.get(array_ref)?.fields.len();
-    if len == 0 {
-        return Ok(Some(Slot::Int(0)));
-    }
-
-    let mut count = 0_usize;
-    for idx in 0..len {
-        let next = heap.read_host_file_byte(file_id)?;
-        if next < 0 {
-            break;
-        }
-        heap.get_mut(array_ref)?.fields[idx] = Slot::Int(next);
-        count += 1;
-    }
-
-    if count == 0 {
-        Ok(Some(Slot::Int(-1)))
-    } else {
-        Ok(Some(Slot::Int(i32::try_from(count).unwrap_or(i32::MAX))))
-    }
-}
-
-fn native_file_input_stream_close(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let file_id = match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(id)) => *id,
-        _ => {
-            return Err(VmError::JavaException {
-                class_name: "java/io/IOException".to_string(),
-            });
-        }
-    };
-    heap.close_host_file(file_id);
-    let stream_obj = heap.get_mut(this_ref)?;
-    let Some(fd_field) = stream_obj.fields.first_mut() else {
-        return Err(VmError::InvalidRef { address: this_ref });
-    };
-    *fd_field = Slot::Int(0);
-    Ok(None)
-}
-
-fn native_file_output_stream_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let path = path_from_string_slot(args, 1, heap)?;
-    let file_id = heap.open_host_output_file(&path)?;
-    let stream_obj = heap.get_mut(this_ref)?;
-    let Some(fd_field) = stream_obj.fields.first_mut() else {
-        return Err(VmError::InvalidRef { address: this_ref });
-    };
-    *fd_field = Slot::Int(file_id);
-    Ok(None)
-}
-
-fn native_file_output_stream_write(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let file_id = file_stream_id_from_this(args, heap)?;
-    let value = match args.get(1) {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    heap.write_host_file_byte(file_id, value)?;
-    Ok(None)
-}
-
-fn native_file_output_stream_write_bytes(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let file_id = file_stream_id_from_this(args, heap)?;
-    let array_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let bytes = heap.get(array_ref)?.fields.clone();
-    for byte in bytes {
-        let Slot::Int(value) = byte else {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        };
-        heap.write_host_file_byte(file_id, value)?;
-    }
-    Ok(None)
-}
-
-fn native_file_output_stream_close(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let file_id = match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(id)) => *id,
-        _ => {
-            return Err(VmError::JavaException {
-                class_name: "java/io/IOException".to_string(),
-            });
-        }
-    };
-    heap.close_host_file(file_id);
-    let stream_obj = heap.get_mut(this_ref)?;
-    let Some(fd_field) = stream_obj.fields.first_mut() else {
-        return Err(VmError::InvalidRef { address: this_ref });
-    };
-    *fd_field = Slot::Int(0);
-    Ok(None)
-}
-
-// ── Networking natives ────────────────────────────────────────────────────
-
-/// Native: `ServerSocket.<init>(int port)` — binds to 0.0.0.0:{port}.
-fn native_server_socket_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let port = match args.get(1) {
-        Some(Slot::Int(p)) => *p,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let addr = format!("0.0.0.0:{port}");
-    let server_id = heap.bind_server_socket(&addr)?;
-    let actual_port = heap.server_socket_local_port(server_id)?;
-    let obj = heap.get_mut(this_ref)?;
-    if obj.fields.len() < 2 {
-        return Err(VmError::InvalidRef { address: this_ref });
-    }
-    obj.fields[0] = Slot::Int(server_id);
-    obj.fields[1] = Slot::Int(actual_port);
-    Ok(None)
-}
-
-/// Native: `ServerSocket.accept()` — blocks until a client connects, returns a Socket.
-fn native_server_socket_accept(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let server_fd = match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(id)) if *id > 0 => *id,
-        _ => {
-            return Err(VmError::JavaException {
-                class_name: "java/io/IOException".into(),
-            });
-        }
-    };
-    let (reader_id, writer_id) = heap.accept_connection(server_fd)?;
-    // Allocate a new Socket object with fdRead=reader_id, fdWrite=writer_id
-    let socket_ref = heap.allocate("java/net/Socket".to_string(), 2);
-    heap.get_mut(socket_ref)?.fields[0] = Slot::Int(reader_id);
-    heap.get_mut(socket_ref)?.fields[1] = Slot::Int(writer_id);
-    Ok(Some(Slot::Reference(Some(socket_ref))))
-}
-
-/// Native: `ServerSocket.getLocalPort()` — returns the bound port.
-fn native_server_socket_get_local_port(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    match heap.get(this_ref)?.fields.get(1) {
-        Some(Slot::Int(port)) => Ok(Some(Slot::Int(*port))),
-        _ => Err(VmError::JavaException {
-            class_name: "java/io/IOException".into(),
-        }),
-    }
-}
-
-/// Native: `ServerSocket.close()` — closes the OS listener and zeros the fd field.
-fn native_server_socket_close(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let fd = match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(id)) if *id > 0 => *id,
-        Some(Slot::Int(_)) => return Ok(None), // already closed — idempotent
-        _ => {
-            return Err(VmError::JavaException {
-                class_name: "java/io/IOException".into(),
-            });
-        }
-    };
-    heap.close_host_file(fd);
-    let obj = heap.get_mut(this_ref)?;
-    obj.fields[0] = Slot::Int(0);
-    obj.fields[1] = Slot::Int(0); // also zero cached port so getLocalPort() returns 0 after close
-    Ok(None)
-}
-
-/// Native: `Socket.<init>(String host, int port)` — connects to host:port.
-fn native_socket_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let host = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => heap
-            .get(*r)?
-            .string_value
-            .clone()
-            .ok_or(VmError::NullPointerException)?,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let port = match args.get(2) {
-        Some(Slot::Int(p)) => *p,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let addr = format!("{host}:{port}");
-    let (reader_id, writer_id) = heap.connect_socket(&addr)?;
-    let obj = heap.get_mut(this_ref)?;
-    if obj.fields.len() < 2 {
-        return Err(VmError::InvalidRef { address: this_ref });
-    }
-    obj.fields[0] = Slot::Int(reader_id);
-    obj.fields[1] = Slot::Int(writer_id);
-    Ok(None)
-}
-
-/// Native: `Socket.getInputStream()` — allocates a `SocketInputStream` wrapping `fdRead`.
-fn native_socket_get_input_stream(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let fd_read = match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(id)) if *id > 0 => *id,
-        _ => {
-            return Err(VmError::JavaException {
-                class_name: "java/io/IOException".into(),
-            });
-        }
-    };
-    let stream_ref = heap.allocate("duke/net/SocketInputStream".to_string(), 1);
-    heap.get_mut(stream_ref)?.fields[0] = Slot::Int(fd_read);
-    Ok(Some(Slot::Reference(Some(stream_ref))))
-}
-
-/// Native: `Socket.getOutputStream()` — allocates a `SocketOutputStream` wrapping `fdWrite`.
-fn native_socket_get_output_stream(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let fd_write = match heap.get(this_ref)?.fields.get(1) {
-        Some(Slot::Int(id)) if *id > 0 => *id,
-        _ => {
-            return Err(VmError::JavaException {
-                class_name: "java/io/IOException".into(),
-            });
-        }
-    };
-    let stream_ref = heap.allocate("duke/net/SocketOutputStream".to_string(), 1);
-    heap.get_mut(stream_ref)?.fields[0] = Slot::Int(fd_write);
-    Ok(Some(Slot::Reference(Some(stream_ref))))
-}
-
-/// Native: `Socket.close()` — closes both OS handles (fdRead and fdWrite).
-fn native_socket_close(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let fd_read = match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(id)) => *id,
-        _ => {
-            return Err(VmError::JavaException {
-                class_name: "java/io/IOException".into(),
-            });
-        }
-    };
-    let fd_write = match heap.get(this_ref)?.fields.get(1) {
-        Some(Slot::Int(id)) => *id,
-        _ => {
-            return Err(VmError::JavaException {
-                class_name: "java/io/IOException".into(),
-            });
-        }
-    };
-    heap.close_host_file(fd_read);
-    heap.close_host_file(fd_write);
-    let obj = heap.get_mut(this_ref)?;
-    obj.fields[0] = Slot::Int(0);
-    obj.fields[1] = Slot::Int(0);
-    Ok(None)
-}
-
-// ── ZIP / JAR natives ──────────────────────────────────────────────────
-
-/// Native: `ZipFile.<init>(String)` — open and index a ZIP/JAR archive.
-fn native_zip_file_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let path_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let path_str = heap
-        .get(path_ref)?
-        .string_value
-        .as_deref()
-        .ok_or(VmError::NullPointerException)?
-        .to_string();
-    let fd = heap.open_host_zip(std::path::Path::new(&path_str))?;
-    let obj = heap.get_mut(this_ref)?;
-    obj.fields[0] = Slot::Int(fd);
-    Ok(None)
-}
-
-/// Native: `ZipFile.getEntry(String) -> ZipEntry` — look up an entry by name.
-#[allow(clippy::cast_possible_truncation, clippy::cast_possible_wrap)]
-fn native_zip_file_get_entry(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let name_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let fd = match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(id)) => *id,
-        _ => {
-            return Err(VmError::JavaException {
-                class_name: "java/io/IOException".into(),
-            });
-        }
-    };
-    let entry_name = heap
-        .get(name_ref)?
-        .string_value
-        .as_deref()
-        .ok_or(VmError::NullPointerException)?
-        .to_string();
-    let info = heap.zip_get_entry_info(fd, &entry_name)?;
-    let Some(info) = info else {
-        return Ok(Some(Slot::Reference(None)));
-    };
-    // Allocate a ZipEntry HeapObject with 6 fields.
-    let name_heap_ref = heap.allocate_string(info.name);
-    let entry_ref = heap.allocate("java/util/zip/ZipEntry".to_string(), 6);
-    let entry_obj = heap.get_mut(entry_ref)?;
-    entry_obj.fields[0] = Slot::Reference(Some(name_heap_ref));
-    entry_obj.fields[1] = Slot::Int(info.compressed_size as i32);
-    entry_obj.fields[2] = Slot::Int((info.compressed_size >> 32) as i32);
-    entry_obj.fields[3] = Slot::Int(info.uncompressed_size as i32);
-    entry_obj.fields[4] = Slot::Int((info.uncompressed_size >> 32) as i32);
-    entry_obj.fields[5] = Slot::Int(i32::from(info.compression_method));
-    Ok(Some(Slot::Reference(Some(entry_ref))))
-}
-
-/// Native: `ZipFile.getInputStream(ZipEntry) -> InputStream`
-fn native_zip_file_get_input_stream(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let entry_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let fd = match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(id)) => *id,
-        _ => {
-            return Err(VmError::JavaException {
-                class_name: "java/io/IOException".into(),
-            });
-        }
-    };
-    // Get entry name from the ZipEntry object.
-    let name_slot_ref = match heap.get(entry_ref)?.fields.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let entry_name = heap
-        .get(name_slot_ref)?
-        .string_value
-        .as_deref()
-        .ok_or(VmError::NullPointerException)?
-        .to_string();
-    // Decompress the entry and wrap in a ByteBuffer.
-    let data = heap.zip_read_entry(fd, &entry_name)?;
-    let buf_fd = heap.open_host_byte_buffer(data);
-    let is_ref = heap.allocate("duke/zip/ByteBufferInputStream".to_string(), 1);
-    let is_obj = heap.get_mut(is_ref)?;
-    is_obj.fields[0] = Slot::Int(buf_fd);
-    Ok(Some(Slot::Reference(Some(is_ref))))
-}
-
-/// Native: `ZipFile.close()`
-fn native_zip_file_close(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let fd = match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(id)) => *id,
-        _ => return Ok(None),
-    };
-    heap.close_host_file(fd);
-    let obj = heap.get_mut(this_ref)?;
-    obj.fields[0] = Slot::Int(0);
-    Ok(None)
-}
-
-/// Native: `ZipFile.size() -> int`
-#[allow(clippy::cast_possible_truncation, clippy::cast_possible_wrap)]
-fn native_zip_file_size(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let fd = match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(id)) => *id,
-        _ => {
-            return Err(VmError::JavaException {
-                class_name: "java/io/IOException".into(),
-            });
-        }
-    };
-    let count = heap.zip_entry_count(fd)?;
-    Ok(Some(Slot::Int(count as i32)))
-}
-
-/// Native: `ZipEntry.getName() -> String`
-fn native_zip_entry_get_name(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    Ok(Some(heap.get(this_ref)?.fields[0]))
-}
-
-/// Native: `ZipEntry.getCompressedSize() -> long`
-fn native_zip_entry_get_compressed_size(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let fields = &heap.get(this_ref)?.fields;
-    let lo = match fields[1] {
-        Slot::Int(v) => v,
-        _ => 0,
-    };
-    let hi = match fields[2] {
-        Slot::Int(v) => v,
-        _ => 0,
-    };
-    let val = (i64::from(hi) << 32) | (i64::from(lo) & 0xFFFF_FFFF);
-    Ok(Some(Slot::Long(val)))
-}
-
-/// Native: `ZipEntry.getSize() -> long`
-fn native_zip_entry_get_size(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let fields = &heap.get(this_ref)?.fields;
-    let lo = match fields[3] {
-        Slot::Int(v) => v,
-        _ => 0,
-    };
-    let hi = match fields[4] {
-        Slot::Int(v) => v,
-        _ => 0,
-    };
-    let val = (i64::from(hi) << 32) | (i64::from(lo) & 0xFFFF_FFFF);
-    Ok(Some(Slot::Long(val)))
-}
-
-/// Native: `ZipEntry.getMethod() -> int`
-fn native_zip_entry_get_method(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    Ok(Some(heap.get(this_ref)?.fields[5]))
-}
-
-/// Native: `String.length()` — returns string length as int.
-#[allow(clippy::cast_possible_truncation, clippy::cast_possible_wrap)]
-fn native_string_length(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let obj = heap.get(this_ref)?;
-    let len = obj.string_value.as_ref().map_or(0, String::len);
-    Ok(Some(Slot::Int(len as i32)))
-}
-
-/// Native: `String.equals(Object)` — compares string content.
-fn native_string_equals(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let other_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Ok(Some(Slot::Int(0))),
-    };
-    // Fetch objects from heap in one go to keep borrows short
-    let this_obj = heap.get(this_ref)?;
-    let other_obj = heap.get(other_ref)?;
-    let this_str = this_obj.string_value.as_deref().unwrap_or_default();
-    let other_str = other_obj.string_value.as_deref().unwrap_or_default();
-    Ok(Some(Slot::Int(i32::from(this_str == other_str))))
-}
-
-/// Native: `String.charAt(int)` — returns char at index as int.
-#[allow(clippy::cast_sign_loss)]
-fn native_string_char_at(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let index = match args.get(1) {
-        Some(Slot::Int(i)) => *i,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let obj = heap.get(this_ref)?;
-    let s = obj.string_value.as_deref().unwrap_or("");
-    let ch = s
-        .chars()
-        .nth(index as usize)
-        .ok_or(VmError::ArrayIndexOutOfBounds {
-            index,
-            length: s.len(),
-        })?;
-    Ok(Some(Slot::Int(ch as i32)))
-}
-
-/// Native: `Object.hashCode()` — returns heap address as hash.
-fn native_object_hashcode(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    match args.first() {
-        #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-        Some(Slot::Reference(Some(r))) => Ok(Some(Slot::Int(*r as i32))),
-        _ => Err(VmError::NullPointerException),
-    }
-}
-
-/// Native: `Object.toString()` — delegates to `heap_object_to_string` so String,
-/// boxed primitives, and opaque objects all produce the correct Java representation.
-fn native_object_tostring(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s = heap_object_to_string(heap.get(this_ref)?, this_ref);
-    let r = heap.allocate_string(s);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `Object.clone()` — shallow-copies a heap object.
-fn native_object_clone(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let obj = heap.get(this_ref)?;
-    let cloned_class = obj.class_name.clone();
-    let cloned_fields = obj.fields.clone();
-    let cloned_string = obj.string_value.clone();
-    let new_ref = heap.allocate(cloned_class, 0);
-    let dest = heap.get_mut(new_ref)?;
-    dest.fields = cloned_fields;
-    dest.string_value = cloned_string;
-    Ok(Some(Slot::Reference(Some(new_ref))))
-}
-
-/// `Throwable.addSuppressed(Throwable suppressed)V`
-///
-/// No-op stub. Control flow is handled entirely by the bytecode desugaring —
-/// `addSuppressed` only affects what `getSuppressed()` returns, which is not
-/// yet implemented. Suppressed exception is silently dropped.
-///
-/// Signature: `args[0]` = this (Throwable), `args[1]` = suppressed (Throwable)
-#[allow(clippy::unnecessary_wraps)]
-fn native_throwable_add_suppressed(
-    _args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _stdout: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    Ok(None)
-}
-
-/// Native: `Enum.<init>(Ljava/lang/String;I)V` — stores name + ordinal.
-/// args: `[this_ref, name_ref, ordinal_int]`
-fn native_enum_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let name_slot = args.get(1).copied().unwrap_or(Slot::Reference(None));
-    let ordinal = match args.get(2) {
-        Some(Slot::Int(v)) => *v,
-        _ => 0,
-    };
-    let obj = heap.get_mut(this_ref)?;
-    if obj.fields.len() >= 2 {
-        obj.fields[0] = name_slot;
-        obj.fields[1] = Slot::Int(ordinal);
-    }
-    Ok(None)
-}
-
-/// Native: `Enum.ordinal()I`
-fn native_enum_ordinal(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let obj = heap.get(this_ref)?;
-    match obj.fields.get(1) {
-        Some(Slot::Int(v)) => Ok(Some(Slot::Int(*v))),
-        _ => Ok(Some(Slot::Int(0))),
-    }
-}
-
-/// Native: `Enum.name()Ljava/lang/String;`
-fn native_enum_name(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let obj = heap.get(this_ref)?;
-    match obj.fields.first() {
-        Some(slot @ Slot::Reference(_)) => Ok(Some(*slot)),
-        _ => Ok(Some(Slot::Reference(None))),
-    }
-}
-
-/// Native: `Enum.valueOf(Ljava/lang/Class;Ljava/lang/String;)Ljava/lang/Enum;`
-/// Searches heap for enum constants of the given class matching the name.
-#[allow(clippy::cast_possible_truncation)]
-fn native_enum_valueof(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let class_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let name_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let target_name = heap.get(name_ref)?.string_value.clone().unwrap_or_default();
-    let enum_class_name = heap
-        .get(class_ref)?
-        .string_value
-        .clone()
-        .unwrap_or_default();
-
-    let obj_count = heap.len();
-    for i in 0..obj_count {
-        let obj = heap.get(i as u64)?;
-        if obj.class_name == enum_class_name
-            && obj.fields.len() >= 2
-            && let Some(Slot::Reference(Some(name_r))) = obj.fields.first()
-            && let Ok(name_obj) = heap.get(*name_r)
-            && name_obj.string_value.as_deref() == Some(target_name.as_str())
-        {
-            return Ok(Some(Slot::Reference(Some(i as u64))));
-        }
-    }
-
-    Err(VmError::JavaException {
-        class_name: "java/lang/IllegalArgumentException".to_string(),
-    })
-}
-
-/// Native: `String.valueOf(int)` — static method, returns string of int.
-fn native_string_value_of_int(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let s = val.to_string();
-    let r = heap.allocate_string(s);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `PrintStream.print(String)` — no newline.
-fn native_print_string(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let string_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => {
-            write!(out, "null").ok();
-            return Ok(None);
-        }
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    let obj = heap.get(string_ref)?;
-    let text = obj.string_value.as_deref().unwrap_or("null");
-    write!(out, "{text}").ok();
-    Ok(None)
-}
-
-/// Native: `PrintStream.print(int)` — no newline.
-#[allow(clippy::unnecessary_wraps)]
-fn native_print_int(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    write!(out, "{val}").ok();
-    Ok(None)
-}
-
-// ---------------------------------------------------------------------------
-// println overloads (long, float, double, boolean, char, object)
-// ---------------------------------------------------------------------------
-
-fn native_println_long(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Long(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Long",
-                got: "other",
-            });
-        }
-    };
-    writeln!(out, "{val}").ok();
-    Ok(None)
-}
-
-fn native_println_float(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Float(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Float",
-                got: "other",
-            });
-        }
-    };
-    writeln!(out, "{val}").ok();
-    Ok(None)
-}
-
-fn native_println_double(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    writeln!(out, "{val}").ok();
-    Ok(None)
-}
-
-fn native_println_boolean(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Int(v)) => *v != 0,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int(boolean)",
-                got: "other",
-            });
-        }
-    };
-    writeln!(out, "{val}").ok();
-    Ok(None)
-}
-
-fn native_println_char(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Int(v)) => char::from_u32((*v).cast_unsigned()).unwrap_or('?'),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int(char)",
-                got: "other",
-            });
-        }
-    };
-    writeln!(out, "{val}").ok();
-    Ok(None)
-}
-
-#[allow(clippy::cast_possible_wrap)]
-/// Convert a heap object to its Java display string.
-///
-/// Checks `string_value` first (handles String/StringBuilder).
-/// For boxed primitives, extracts the stored value from `fields[0]`.
-/// Falls back to `class_name@hex_ref` for opaque objects.
-fn heap_object_to_string(obj: &duke_gc::HeapObject, obj_ref: u64) -> String {
-    if let Some(s) = &obj.string_value {
-        return s.clone();
-    }
-    match obj.class_name.as_str() {
-        "java/lang/Integer" => {
-            if let Some(Slot::Int(v)) = obj.fields.first() {
-                return v.to_string();
-            }
-        }
-        "java/lang/Long" => {
-            if let Some(Slot::Long(v)) = obj.fields.first() {
-                return v.to_string();
-            }
-        }
-        "java/lang/Double" => {
-            if let Some(Slot::Double(v)) = obj.fields.first() {
-                return format_java_double(*v);
-            }
-        }
-        "java/lang/Float" => {
-            if let Some(Slot::Float(v)) = obj.fields.first() {
-                return format_java_float(*v);
-            }
-        }
-        "java/lang/Boolean" => {
-            return match obj.fields.first() {
-                Some(Slot::Int(v)) if *v != 0 => "true".to_string(),
-                _ => "false".to_string(),
-            };
-        }
-        "java/lang/Character" => {
-            if let Some(Slot::Int(v)) = obj.fields.first()
-                && let Some(c) = char::from_u32((*v).cast_unsigned())
-            {
-                return c.to_string();
-            }
-        }
-        _ => {}
-    }
-    format!("{}@{:x}", obj.class_name, obj_ref)
-}
-
-fn native_println_object(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    match args.get(1) {
-        Some(Slot::Reference(Some(r))) => {
-            let s = heap_object_to_string(heap.get(*r)?, *r);
-            writeln!(out, "{s}").ok();
-        }
-        Some(Slot::Reference(None)) => {
-            writeln!(out, "null").ok();
-        }
-        _ => {
-            writeln!(out, "<unknown>").ok();
-        }
-    }
-    Ok(None)
-}
-
-// ---------------------------------------------------------------------------
-// print overloads (long, float, double, boolean, char, object)
-// ---------------------------------------------------------------------------
-
-fn native_print_long(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Long(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Long",
-                got: "other",
-            });
-        }
-    };
-    write!(out, "{val}").ok();
-    Ok(None)
-}
-
-fn native_print_float(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Float(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Float",
-                got: "other",
-            });
-        }
-    };
-    write!(out, "{val}").ok();
-    Ok(None)
-}
-
-fn native_print_double(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    write!(out, "{val}").ok();
-    Ok(None)
-}
-
-fn native_print_boolean(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Int(v)) => *v != 0,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int(boolean)",
-                got: "other",
-            });
-        }
-    };
-    write!(out, "{val}").ok();
-    Ok(None)
-}
-
-fn native_print_char(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.get(1) {
-        Some(Slot::Int(v)) => char::from_u32((*v).cast_unsigned()).unwrap_or('?'),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int(char)",
-                got: "other",
-            });
-        }
-    };
-    write!(out, "{val}").ok();
-    Ok(None)
-}
-
-#[allow(clippy::cast_possible_truncation, clippy::cast_possible_wrap)]
-fn native_print_object(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    match args.get(1) {
-        Some(Slot::Reference(Some(r))) => {
-            let s = heap_object_to_string(heap.get(*r)?, *r);
-            write!(out, "{s}").ok();
-        }
-        Some(Slot::Reference(None)) => {
-            write!(out, "null").ok();
-        }
-        _ => {
-            write!(out, "<unknown>").ok();
-        }
-    }
-    Ok(None)
-}
-
-// ---------------------------------------------------------------------------
-// System.exit
-// ---------------------------------------------------------------------------
-
-fn native_system_exit(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let code = match args.first() {
-        Some(Slot::Int(v)) => *v,
-        _ => 1,
-    };
-    Err(VmError::SystemExit { code })
-}
-
-const THREAD_TARGET_SLOT: usize = 0;
-const THREAD_ID_SLOT: usize = 1;
-
-fn native_thread_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let this = heap.get_mut(this_ref)?;
-    this.fields[THREAD_TARGET_SLOT] = Slot::Reference(None);
-    this.fields[THREAD_ID_SLOT] = Slot::Int(-1);
-    Ok(None)
-}
-
-fn native_thread_init_runnable(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let target = args.get(1).copied().unwrap_or(Slot::Reference(None));
-    let this = heap.get_mut(this_ref)?;
-    this.fields[THREAD_TARGET_SLOT] = target;
-    this.fields[THREAD_ID_SLOT] = Slot::Int(-1);
-    Ok(None)
-}
-
-fn native_thread_start(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let thread_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    control.request(NativeThreadAction::Start { thread_ref });
-    Ok(None)
-}
-
-fn native_thread_join(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let thread_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let thread = heap.get(thread_ref)?;
-    let Some(Slot::Int(thread_id)) = thread.fields.get(THREAD_ID_SLOT) else {
-        return Ok(None);
-    };
-    if *thread_id >= 0 {
-        control.request(NativeThreadAction::Join {
-            thread_id: *thread_id,
-        });
-    }
-    Ok(None)
-}
-
-fn native_thread_sleep(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let millis = match args.first() {
-        Some(Slot::Long(value)) => *value,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "long",
-                got: "other",
-            });
-        }
-    };
-    let millis = u64::try_from(millis.max(0)).unwrap_or(0);
-    control.request(NativeThreadAction::Sleep(std::time::Duration::from_millis(
-        millis,
-    )));
-    Ok(None)
-}
-
-/// Native: `String.substring(int)` — substring from begin to end.
-#[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-fn native_string_substring(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-
-    let begin = match args.get(1) {
-        Some(Slot::Int(v)) => *v as usize,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let sub = {
-        let obj = heap.get(this_ref)?;
-        let s = obj.string_value.as_deref().unwrap_or_default();
-        if begin > s.len() {
-            return Err(VmError::ArrayIndexOutOfBounds {
-                index: i32::try_from(begin).unwrap_or(i32::MAX),
-                length: s.len(),
-            });
-        }
-        s.chars().skip(begin).collect::<String>()
-    };
-
-    let r = heap.allocate_string(sub);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.substring(int, int)` — substring from begin to end (exclusive).
-#[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-fn native_string_substring_range(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-
-    let begin = match args.get(1) {
-        Some(Slot::Int(v)) => *v as usize,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let end = match args.get(2) {
-        Some(Slot::Int(v)) => *v as usize,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let sub = {
-        let obj = heap.get(this_ref)?;
-        let s = obj.string_value.as_deref().unwrap_or_default();
-        if begin > end || end > s.len() {
-            return Err(VmError::ArrayIndexOutOfBounds {
-                index: i32::try_from(end).unwrap_or(i32::MAX),
-                length: s.len(),
-            });
-        }
-        s.chars().skip(begin).take(end - begin).collect::<String>()
-    };
-
-    let r = heap.allocate_string(sub);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.indexOf(String)` — find first occurrence of target.
-fn native_string_indexof(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-
-    let target_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    // Fetch objects from heap in one go to keep borrows short
-    let this_obj = heap.get(this_ref)?;
-    let target_obj = heap.get(target_ref)?;
-    let s = this_obj.string_value.as_deref().unwrap_or_default();
-    let target = target_obj.string_value.as_deref().unwrap_or_default();
-
-    #[allow(clippy::cast_possible_truncation, clippy::cast_possible_wrap)]
-    let result = s.find(target).map_or(-1, |i| i as i32);
-    Ok(Some(Slot::Int(result)))
-}
-
-/// Native: `String.contains(CharSequence)` — check if string contains target.
-fn native_string_contains(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-
-    let target_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    // Fetch objects from heap in one go to keep borrows short
-    let this_obj = heap.get(this_ref)?;
-    let target_obj = heap.get(target_ref)?;
-    let s = this_obj.string_value.as_deref().unwrap_or_default();
-    let target = target_obj.string_value.as_deref().unwrap_or_default();
-    Ok(Some(Slot::Int(i32::from(s.contains(target)))))
-}
-
-/// Native: `String.isEmpty()` — check if string is empty.
-#[allow(clippy::unnecessary_wraps)]
-fn native_string_isempty(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let obj = heap.get(this_ref)?;
-    let s = obj.string_value.as_deref().unwrap_or_default();
-    Ok(Some(Slot::Int(i32::from(s.is_empty()))))
-}
-
-// Dispatch string constants used by the callback-based sort chain.
-const COMPARE_TO_METHOD: &str = "compareTo";
-const COMPARE_TO_OBJECT_DESC: &str = "(Ljava/lang/Object;)I";
-const SORT_COMPARATOR_DESC: &str = "(Ljava/util/Comparator;)V";
-
-/// Maps a `std::cmp::Ordering` to the Java `compareTo` convention: -1 / 0 / 1.
-///
-/// Used by all boxed-type `compareTo` natives to return a consistent,
-/// sign-correct value without relying on `Ordering`'s internal discriminant.
-#[inline]
-const fn ordering_to_int(o: std::cmp::Ordering) -> i32 {
-    match o {
-        std::cmp::Ordering::Less => -1,
-        std::cmp::Ordering::Equal => 0,
-        std::cmp::Ordering::Greater => 1,
-    }
-}
-
-/// Native: `String.compareTo(String)` — delegates to the Object overload.
-fn native_string_compareto(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    out: &mut dyn Write,
-    control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    native_string_compareto_object(args, heap, out, control)
-}
-
-/// Native: `String.compareTo(Object)` — lexicographic comparison via Object descriptor.
-fn native_string_compareto_object(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let str_val = |s: &Slot| -> VmResult<String> {
-        match s {
-            Slot::Reference(Some(r)) => Ok(heap.get(*r)?.string_value.clone().unwrap_or_default()),
-            _ => Err(VmError::NullPointerException),
-        }
-    };
-    let a = match args.first() {
-        Some(s) => str_val(s)?,
-        None => return Err(VmError::NullPointerException),
-    };
-    let b = match args.get(1) {
-        Some(s) => str_val(s)?,
-        None => return Err(VmError::NullPointerException),
-    };
-    Ok(Some(Slot::Int(ordering_to_int(a.as_str().cmp(b.as_str())))))
-}
-
-/// Native: `String.startsWith(String)` — check if string starts with prefix.
-fn native_string_startswith(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let prefix_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let prefix = heap
-        .get(prefix_ref)?
-        .string_value
-        .clone()
-        .unwrap_or_default();
-    Ok(Some(Slot::Int(i32::from(s.starts_with(&prefix)))))
-}
-
-/// Native: `String.endsWith(String)` — check if string ends with suffix.
-fn native_string_endswith(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let suffix_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let suffix = heap
-        .get(suffix_ref)?
-        .string_value
-        .clone()
-        .unwrap_or_default();
-    Ok(Some(Slot::Int(i32::from(s.ends_with(&suffix)))))
-}
-
-/// Native: `String.trim()` — remove leading and trailing whitespace.
-fn native_string_trim(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let trimmed = s.trim().to_string();
-    let r = heap.allocate_string(trimmed);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.toCharArray()` — convert string to char array.
-#[allow(clippy::cast_possible_truncation, clippy::cast_possible_wrap)]
-fn native_string_tochararray(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let chars: Vec<char> = s.chars().collect();
-    let arr_ref = heap.allocate("[C".to_string(), chars.len());
-    for (i, &c) in chars.iter().enumerate() {
-        heap.get_mut(arr_ref).unwrap().fields[i] = Slot::Int(c as i32);
-    }
-    Ok(Some(Slot::Reference(Some(arr_ref))))
-}
-
-// ---- Integer natives ----
-
-/// Native: `Integer.parseInt(String)` — parses string to int.
-fn native_integer_parseint(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let str_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    let s = heap.get(str_ref)?.string_value.clone().unwrap_or_default();
-    let val: i32 = s.trim().parse().map_err(|_| VmError::JavaException {
-        class_name: "java/lang/NumberFormatException".to_string(),
-    })?;
-    Ok(Some(Slot::Int(val)))
-}
-
-/// Native: `Integer.valueOf(int)` — boxes int into Integer object.
-fn native_integer_valueof(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let r = heap.allocate("java/lang/Integer".to_string(), 1);
-    heap.get_mut(r).unwrap().fields[0] = Slot::Int(val);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `Integer.intValue()` — unboxes Integer to int.
-fn native_integer_intvalue(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let val = heap.get(this_ref)?.fields[0];
-    Ok(Some(val))
-}
-
-/// Native: `Integer.toString(int)` — static, converts int to String.
-fn native_integer_tostring_static(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let r = heap.allocate_string(val.to_string());
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `Integer.compareTo(Object)` — compares two boxed Integers.
-fn native_integer_compareto(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let int_val = |s: &Slot| -> VmResult<i32> {
-        match s {
-            Slot::Reference(Some(r)) => match heap.get(*r)?.fields.first() {
-                Some(Slot::Int(n)) => Ok(*n),
-                _ => Err(VmError::InvalidRef { address: *r }),
-            },
-            _ => Err(VmError::NullPointerException),
-        }
-    };
-    let a = match args.first() {
-        Some(s) => int_val(s)?,
-        None => return Err(VmError::NullPointerException),
-    };
-    let b = match args.get(1) {
-        Some(s) => int_val(s)?,
-        None => return Err(VmError::NullPointerException),
-    };
-    Ok(Some(Slot::Int(ordering_to_int(a.cmp(&b)))))
-}
-
-// ---- String.valueOf overloads ----
-
-/// Native: `String.valueOf(long)` — converts long to String.
-fn native_string_value_of_long(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Long(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Long",
-                got: "other",
-            });
-        }
-    };
-    let r = heap.allocate_string(val.to_string());
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.valueOf(double)` — converts double to String.
-fn native_string_value_of_double(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    let r = heap.allocate_string(val.to_string());
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.valueOf(float)` — converts float to String.
-fn native_string_value_of_float(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Float(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Float",
-                got: "other",
-            });
-        }
-    };
-    let r = heap.allocate_string(val.to_string());
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.valueOf(boolean)` — converts boolean to String.
-fn native_string_value_of_boolean(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Int(v)) => *v != 0,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int(boolean)",
-                got: "other",
-            });
-        }
-    };
-    let r = heap.allocate_string(if val { "true" } else { "false" }.to_string());
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.valueOf(char)` — converts char to String.
-fn native_string_value_of_char(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Int(v)) => char::from_u32((*v).cast_unsigned()).unwrap_or('?'),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int(char)",
-                got: "other",
-            });
-        }
-    };
-    let r = heap.allocate_string(val.to_string());
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.valueOf(Object)` — converts Object to String.
-fn native_string_value_of_object(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    match args.first() {
-        Some(Slot::Reference(Some(r))) => {
-            let s = heap_object_to_string(heap.get(*r)?, *r);
-            let r = heap.allocate_string(s);
-            Ok(Some(Slot::Reference(Some(r))))
-        }
-        Some(Slot::Reference(None)) => {
-            let r = heap.allocate_string("null".to_string());
-            Ok(Some(Slot::Reference(Some(r))))
-        }
-        _ => Err(VmError::TypeMismatch {
-            expected: "Reference",
-            got: "other",
-        }),
-    }
-}
-
-// ---- String.concat ----
-
-/// Native: `String.concat(String)` — concatenates two strings.
-fn native_string_concat(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s1 = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let other_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    let s2 = heap
-        .get(other_ref)?
-        .string_value
-        .clone()
-        .unwrap_or_default();
-    let r = heap.allocate_string(format!("{s1}{s2}"));
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Formats a single boxed slot value using the given format specifier.
-fn format_arg(
-    spec: char,
-    precision: Option<usize>,
-    slot: &Slot,
-    heap: &duke_gc::Heap,
-) -> VmResult<String> {
-    match slot {
-        Slot::Reference(None) => Ok("null".to_string()),
-        Slot::Reference(Some(r)) => {
-            let obj = heap.get(*r)?;
-            match spec {
-                's' => Ok(heap_object_to_string(obj, *r)),
-                'd' => match obj.fields.first() {
-                    Some(Slot::Int(v)) => Ok(v.to_string()),
-                    Some(Slot::Long(v)) => Ok(v.to_string()),
-                    _ => Ok("0".to_string()),
-                },
-                'f' => {
-                    let v = match obj.fields.first() {
-                        Some(Slot::Double(v)) => *v,
-                        Some(Slot::Float(v)) => f64::from(*v),
-                        _ => 0.0,
-                    };
-                    Ok(precision.map_or_else(|| format!("{v:.6}"), |p| format!("{v:.p$}")))
-                }
-                'x' => match obj.fields.first() {
-                    Some(Slot::Int(v)) => Ok(format!("{v:x}")),
-                    Some(Slot::Long(v)) => Ok(format!("{v:x}")),
-                    _ => Ok("0".to_string()),
-                },
-                'X' => match obj.fields.first() {
-                    Some(Slot::Int(v)) => Ok(format!("{v:X}")),
-                    Some(Slot::Long(v)) => Ok(format!("{v:X}")),
-                    _ => Ok("0".to_string()),
-                },
-                _ => Ok(String::new()),
-            }
-        }
-        _ => Ok(String::new()),
-    }
-}
-
-/// Native: `String.format(Ljava/lang/String;[Ljava/lang/Object;)Ljava/lang/String;`
-fn native_string_format(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let fmt_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let fmt = heap.get(fmt_ref)?.string_value.clone().unwrap_or_default();
-
-    let arr_len = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => heap.get(*r)?.fields.len(),
-        _ => 0,
-    };
-
-    let mut result = String::new();
-    let mut arg_idx = 0usize;
-    let chars: Vec<char> = fmt.chars().collect();
-    let mut i = 0;
-
-    while i < chars.len() {
-        if chars[i] != '%' {
-            result.push(chars[i]);
-            i += 1;
-            continue;
-        }
-        i += 1;
-        if i >= chars.len() {
-            break;
-        }
-
-        // Parse optional precision: %.2f
-        let precision: Option<usize> = if chars[i] == '.' {
-            i += 1;
-            let mut prec_str = String::new();
-            while i < chars.len() && chars[i].is_ascii_digit() {
-                prec_str.push(chars[i]);
-                i += 1;
-            }
-            prec_str.parse().ok()
-        } else {
-            None
-        };
-
-        // Skip optional width digits
-        while i < chars.len() && chars[i].is_ascii_digit() {
-            i += 1;
-        }
-
-        if i >= chars.len() {
-            break;
-        }
-        let spec = chars[i];
-        i += 1;
-
-        match spec {
-            '%' => result.push('%'),
-            'n' => result.push('\n'),
-            's' | 'd' | 'f' | 'x' | 'X' => {
-                let slot = if arg_idx < arr_len {
-                    match args.get(1) {
-                        Some(Slot::Reference(Some(r))) => heap
-                            .get(*r)?
-                            .fields
-                            .get(arg_idx)
-                            .copied()
-                            .unwrap_or(Slot::Reference(None)),
-                        _ => Slot::Reference(None),
-                    }
-                } else {
-                    Slot::Reference(None)
-                };
-                arg_idx += 1;
-                let formatted = format_arg(spec, precision, &slot, heap)?;
-                result.push_str(&formatted);
-            }
-            _ => {
-                result.push('%');
-                result.push(spec);
-            }
-        }
-    }
-
-    let r = heap.allocate_string(result);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-// ---- Extended String natives ----
-
-/// Native: `String.toUpperCase()` — returns a new uppercase String.
-fn native_string_touppercase(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let r = heap.allocate_string(s.to_uppercase());
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.toLowerCase()` — returns a new lowercase String.
-fn native_string_tolowercase(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let r = heap.allocate_string(s.to_lowercase());
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.replace(char, char)` — replaces all occurrences of old char with new char.
-#[allow(clippy::cast_sign_loss)]
-fn native_string_replace_char(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let old_char = match args.get(1) {
-        Some(Slot::Int(v)) => char::from_u32((*v).cast_unsigned()).unwrap_or('?'),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let new_char = match args.get(2) {
-        Some(Slot::Int(v)) => char::from_u32((*v).cast_unsigned()).unwrap_or('?'),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let result = s.replace(old_char, &new_char.to_string());
-    let r = heap.allocate_string(result);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.replace(CharSequence, CharSequence)` — replaces all occurrences of target with replacement.
-fn native_string_replace_charsequence(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let target_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    let target = heap
-        .get(target_ref)?
-        .string_value
-        .clone()
-        .unwrap_or_default();
-    let replacement_ref = match args.get(2) {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    let replacement = heap
-        .get(replacement_ref)?
-        .string_value
-        .clone()
-        .unwrap_or_default();
-    let result = s.replace(&*target, &replacement);
-    let r = heap.allocate_string(result);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `String.split(String)` — splits string by delimiter, returns String array.
-fn native_string_split(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let delim_ref = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    let delim = heap
-        .get(delim_ref)?
-        .string_value
-        .clone()
-        .unwrap_or_default();
-    let parts: Vec<&str> = s.split(&*delim).collect();
-    let arr_ref = heap.allocate("[Ljava/lang/String;".to_string(), parts.len());
-    for (i, part) in parts.iter().enumerate() {
-        let str_ref = heap.allocate_string((*part).to_string());
-        heap.get_mut(arr_ref).unwrap().fields[i] = Slot::Reference(Some(str_ref));
-    }
-    Ok(Some(Slot::Reference(Some(arr_ref))))
-}
-
-/// Native: `String.hashCode()` — Java's hash algorithm: `s[0]*31^(n-1) + s[1]*31^(n-2) + ... + s[n-1]`.
-#[allow(clippy::cast_possible_wrap)]
-fn native_string_hashcode(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let mut h: i32 = 0;
-    for ch in s.chars() {
-        h = h.wrapping_mul(31).wrapping_add(ch as i32);
-    }
-    Ok(Some(Slot::Int(h)))
-}
-
-/// Native: `String.toString()` — identity, returns `this`.
-#[allow(clippy::unnecessary_wraps)]
-fn native_string_tostring(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    Ok(Some(args.first().copied().unwrap_or(Slot::Reference(None))))
-}
-
-// ---- Math natives ----
-
-/// Native: `Math.max(int, int)` — returns the larger value.
-fn native_math_max_int(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let b = match args.get(1) {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Int(a.max(b))))
-}
-
-/// Native: `Math.min(int, int)` — returns the smaller value.
-fn native_math_min_int(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let b = match args.get(1) {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Int(a.min(b))))
-}
-
-/// Native: `Math.abs(int)` — returns absolute value.
-fn native_math_abs_int(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Int(a.wrapping_abs())))
-}
-
-// ---- Extended Math natives ----
-
-/// Native: `Math.sqrt(double)` — returns square root.
-fn native_math_sqrt(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Double(a.sqrt())))
-}
-
-/// Native: `Math.pow(double, double)` — returns a raised to the power b.
-fn native_math_pow(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    let b = match args.get(1) {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Double(a.powf(b))))
-}
-
-/// Native: `Math.floor(double)` — returns floor value.
-fn native_math_floor(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Double(a.floor())))
-}
-
-/// Native: `Math.ceil(double)` — returns ceiling value.
-fn native_math_ceil(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Double(a.ceil())))
-}
-
-/// Native: `Math.round(double)` — returns closest long.
-#[allow(clippy::cast_possible_truncation)]
-fn native_math_round_double(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Long(a.round() as i64)))
-}
-
-/// Native: `Math.abs(long)` — returns absolute value.
-fn native_math_abs_long(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Long(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Long",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Long(a.wrapping_abs())))
-}
-
-/// Native: `Math.abs(double)` — returns absolute value.
-fn native_math_abs_double(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Double(a.abs())))
-}
-
-/// Native: `Math.max(long, long)` — returns the larger value.
-fn native_math_max_long(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Long(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Long",
-                got: "other",
-            });
-        }
-    };
-    let b = match args.get(1) {
-        Some(Slot::Long(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Long",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Long(a.max(b))))
-}
-
-/// Native: `Math.min(long, long)` — returns the smaller value.
-fn native_math_min_long(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Long(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Long",
-                got: "other",
-            });
-        }
-    };
-    let b = match args.get(1) {
-        Some(Slot::Long(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Long",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Long(a.min(b))))
-}
-
-/// Native: `Math.max(double, double)` — returns the larger value.
-fn native_math_max_double(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    let b = match args.get(1) {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Double(a.max(b))))
-}
-
-/// Native: `Math.min(double, double)` — returns the smaller value.
-fn native_math_min_double(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let a = match args.first() {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    let b = match args.get(1) {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    Ok(Some(Slot::Double(a.min(b))))
-}
-
-// ---- Long class natives ----
-
-/// Native: `Long.parseLong(String)` — parses string to long.
-fn native_long_parselong(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let str_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    let s = heap.get(str_ref)?.string_value.clone().unwrap_or_default();
-    let val: i64 = s.trim().parse().map_err(|_| VmError::JavaException {
-        class_name: "java/lang/NumberFormatException".to_string(),
-    })?;
-    Ok(Some(Slot::Long(val)))
-}
-
-/// Native: `Long.valueOf(long)` — boxes long into Long object.
-fn native_long_valueof(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Long(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Long",
-                got: "other",
-            });
-        }
-    };
-    let r = heap.allocate("java/lang/Long".to_string(), 1);
-    heap.get_mut(r).unwrap().fields[0] = Slot::Long(val);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `Long.longValue()` — unboxes Long to long.
-fn native_long_longvalue(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let val = heap.get(this_ref)?.fields[0];
-    Ok(Some(val))
-}
-
-/// Native: `Long.toString(long)` — static, converts long to String.
-fn native_long_tostring_static(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Long(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Long",
-                got: "other",
-            });
-        }
-    };
-    let r = heap.allocate_string(val.to_string());
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `Long.compareTo(Object)` — compares two boxed Longs.
-fn native_long_compareto(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let long_val = |s: &Slot| -> VmResult<i64> {
-        match s {
-            Slot::Reference(Some(r)) => match heap.get(*r)?.fields.first() {
-                Some(Slot::Long(n)) => Ok(*n),
-                _ => Err(VmError::InvalidRef { address: *r }),
-            },
-            _ => Err(VmError::NullPointerException),
-        }
-    };
-    let a = match args.first() {
-        Some(s) => long_val(s)?,
-        None => return Err(VmError::NullPointerException),
-    };
-    let b = match args.get(1) {
-        Some(s) => long_val(s)?,
-        None => return Err(VmError::NullPointerException),
-    };
-    Ok(Some(Slot::Int(ordering_to_int(a.cmp(&b)))))
-}
-
-// ---- Double class natives ----
-
-/// Native: `Double.parseDouble(String)` — parses string to double.
-fn native_double_parsedouble(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let str_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    let s = heap.get(str_ref)?.string_value.clone().unwrap_or_default();
-    let val: f64 = s.trim().parse().map_err(|_| VmError::JavaException {
-        class_name: "java/lang/NumberFormatException".to_string(),
-    })?;
-    Ok(Some(Slot::Double(val)))
-}
-
-/// Native: `Double.valueOf(double)` — boxes double into Double object.
-fn native_double_valueof(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    let r = heap.allocate("java/lang/Double".to_string(), 1);
-    heap.get_mut(r).unwrap().fields[0] = Slot::Double(val);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `Double.doubleValue()` — unboxes Double to double.
-fn native_double_doublevalue(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let val = heap.get(this_ref)?.fields[0];
-    Ok(Some(val))
-}
-
-// ---- Float class native ----
-
-/// Native: `Float.parseFloat(String)` — parses string to float.
-fn native_float_parsefloat(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let str_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Reference",
-                got: "other",
-            });
-        }
-    };
-    let s = heap.get(str_ref)?.string_value.clone().unwrap_or_default();
-    let val: f32 = s.trim().parse().map_err(|_| VmError::JavaException {
-        class_name: "java/lang/NumberFormatException".to_string(),
-    })?;
-    Ok(Some(Slot::Float(val)))
-}
-
-// ---- Boolean class native ----
-
-/// Native: `Boolean.parseBoolean(String)` — case-insensitive "true" → 1, else 0.
-fn native_boolean_parseboolean(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    match args.first() {
-        Some(Slot::Reference(Some(r))) => {
-            let s = heap.get(*r)?.string_value.clone().unwrap_or_default();
-            let val = s.eq_ignore_ascii_case("true");
-            Ok(Some(Slot::Int(i32::from(val))))
-        }
-        Some(Slot::Reference(None)) => Ok(Some(Slot::Int(0))),
-        _ => Err(VmError::TypeMismatch {
-            expected: "Reference",
-            got: "other",
-        }),
-    }
-}
-
-/// Execute a `StringConcatFactory` recipe: walk the recipe string, replacing
-/// `\u{1}` placeholders with stringified dynamic args from the operand stack.
-fn execute_string_concat_recipe(
-    recipe: &str,
-    dynamic_args: &[Slot],
-    arg_types: &[char],
-    constants: &[String],
-    heap: &mut duke_gc::Heap,
-) -> VmResult<Slot> {
-    let mut result = String::new();
-    let mut dyn_idx = 0;
-    let mut const_idx = 0;
-
-    for ch in recipe.chars() {
-        match ch {
-            '\u{1}' => {
-                if dyn_idx < dynamic_args.len() {
-                    let type_hint = arg_types.get(dyn_idx).copied().unwrap_or('I');
-                    stringify_slot(&dynamic_args[dyn_idx], type_hint, heap, &mut result)?;
-                    dyn_idx += 1;
-                }
-            }
-            '\u{2}' => {
-                if const_idx < constants.len() {
-                    result.push_str(&constants[const_idx]);
-                    const_idx += 1;
-                }
-            }
-            other => result.push(other),
-        }
-    }
-
-    let r = heap.allocate_string(result);
-    Ok(Slot::Reference(Some(r)))
-}
-
-/// Convert a Slot to its string representation (like Java's String.valueOf).
-/// `type_hint` is the JVM type descriptor char: 'Z' for boolean, 'I' for int, etc.
-fn stringify_slot(
-    slot: &Slot,
-    type_hint: char,
-    heap: &duke_gc::Heap,
-    out: &mut String,
-) -> VmResult<()> {
-    match slot {
-        Slot::Int(v) => {
-            if type_hint == 'Z' {
-                // JVM boolean: 0 = false, nonzero = true.
-                out.push_str(if *v != 0 { "true" } else { "false" });
-            } else if type_hint == 'C' {
-                // JVM char: render as the Unicode character.
-                if let Some(ch) = char::from_u32((*v).cast_unsigned()) {
-                    out.push(ch);
-                } else {
-                    out.push('?');
-                }
-            } else {
-                out.push_str(&v.to_string());
-            }
-        }
-        Slot::Long(v) => out.push_str(&v.to_string()),
-        Slot::Float(v) => out.push_str(&format_java_float(*v)),
-        Slot::Double(v) => out.push_str(&format_java_double(*v)),
-        Slot::Reference(None) => out.push_str("null"),
-        Slot::Reference(Some(r)) => {
-            out.push_str(&heap_object_to_string(heap.get(*r)?, *r));
-        }
-        Slot::ReturnAddress(v) => out.push_str(&v.to_string()),
-    }
-    Ok(())
-}
-
-/// Format a float like Java's Float.toString.
-fn format_java_float(v: f32) -> String {
-    if v.is_nan() {
-        return "NaN".to_string();
-    }
-    if v.is_infinite() {
-        return if v > 0.0 { "Infinity" } else { "-Infinity" }.to_string();
-    }
-    let s = format!("{v}");
-    if s.contains('.') { s } else { format!("{v}.0") }
-}
-
-/// Format a double like Java's Double.toString.
-fn format_java_double(v: f64) -> String {
-    if v.is_nan() {
-        return "NaN".to_string();
-    }
-    if v.is_infinite() {
-        return if v > 0.0 { "Infinity" } else { "-Infinity" }.to_string();
-    }
-    let s = format!("{v}");
-    if s.contains('.') { s } else { format!("{v}.0") }
 }
 
 /// Execute a decoded JVM instruction stream.
@@ -7844,9 +4917,9 @@ fn run_execution(
                 init_object_fields(registry, heap, r, &target_class);
                 frame.push(Slot::Reference(Some(r)))?;
                 if gc_allowed && heap.should_gc() {
-                    let roots = gather_roots(frame, call_stack, registry);
+                    let roots = natives::gather_roots(frame, call_stack, registry);
                     heap.collect(&roots);
-                    patch_forwarded_slots(frame, call_stack, registry, heap);
+                    natives::patch_forwarded_slots(frame, call_stack, registry, heap);
                 }
             }
 
@@ -8374,9 +5447,9 @@ fn run_execution(
                 }
                 frame.push(Slot::Reference(Some(r)))?;
                 if gc_allowed && heap.should_gc() {
-                    let roots = gather_roots(frame, call_stack, registry);
+                    let roots = natives::gather_roots(frame, call_stack, registry);
                     heap.collect(&roots);
-                    patch_forwarded_slots(frame, call_stack, registry, heap);
+                    natives::patch_forwarded_slots(frame, call_stack, registry, heap);
                 }
             }
             Instruction::Anewarray(cp_idx) => {
@@ -8397,9 +5470,9 @@ fn run_execution(
                 }
                 frame.push(Slot::Reference(Some(r)))?;
                 if gc_allowed && heap.should_gc() {
-                    let roots = gather_roots(frame, call_stack, registry);
+                    let roots = natives::gather_roots(frame, call_stack, registry);
                     heap.collect(&roots);
-                    patch_forwarded_slots(frame, call_stack, registry, heap);
+                    natives::patch_forwarded_slots(frame, call_stack, registry, heap);
                 }
             }
             Instruction::Arraylength => {
@@ -8820,7 +5893,7 @@ fn run_execution(
                         (recipe, consts)
                     };
 
-                    let result = execute_string_concat_recipe(
+                    let result = natives::execute_string_concat_recipe(
                         &recipe,
                         &dynamic_args,
                         &arg_types,
@@ -8898,9 +5971,9 @@ fn run_execution(
 
                     frame.push(Slot::Reference(Some(r)))?;
                     if gc_allowed && heap.should_gc() {
-                        let roots = gather_roots(frame, call_stack, registry);
+                        let roots = natives::gather_roots(frame, call_stack, registry);
                         heap.collect(&roots);
-                        patch_forwarded_slots(frame, call_stack, registry, heap);
+                        natives::patch_forwarded_slots(frame, call_stack, registry, heap);
                     }
                 } else {
                     // Unknown bootstrap method — pop args and push null.
@@ -9436,9 +6509,9 @@ fn run_execution(
                 let r = alloc_multi(heap, &dims, 0, &element_type);
                 frame.push(Slot::Reference(Some(r)))?;
                 if gc_allowed && heap.should_gc() {
-                    let roots = gather_roots(frame, call_stack, registry);
+                    let roots = natives::gather_roots(frame, call_stack, registry);
                     heap.collect(&roots);
-                    patch_forwarded_slots(frame, call_stack, registry, heap);
+                    natives::patch_forwarded_slots(frame, call_stack, registry, heap);
                 }
             }
 
@@ -9501,7 +6574,11 @@ fn resolve_thread_entry(
         )));
     }
 
-    let target_ref = match heap.get(thread_ref)?.fields.get(THREAD_TARGET_SLOT) {
+    let target_ref = match heap
+        .get(thread_ref)?
+        .fields
+        .get(natives::THREAD_TARGET_SLOT)
+    {
         Some(Slot::Reference(Some(target_ref))) => *target_ref,
         _ => return Ok(None),
     };
@@ -9636,7 +6713,7 @@ fn spawn_java_thread(
         let mut shared_guard = shared.lock().unwrap();
         let thread = shared_guard.heap.get_mut(thread_ref)?;
         let already_started = matches!(
-            thread.fields.get(THREAD_ID_SLOT),
+            thread.fields.get(natives::THREAD_ID_SLOT),
             Some(Slot::Int(thread_id)) if *thread_id >= 0
         );
         drop(shared_guard);
@@ -9658,7 +6735,7 @@ fn spawn_java_thread(
         let mut shared_guard = shared.lock().unwrap();
         {
             let thread = shared_guard.heap.get_mut(thread_ref)?;
-            thread.fields[THREAD_ID_SLOT] = Slot::Int(thread_id);
+            thread.fields[natives::THREAD_ID_SLOT] = Slot::Int(thread_id);
         }
         let CompletionVm {
             registry,
@@ -10529,1296 +7606,6 @@ fn static_field_idx(ctx: &ClassContext, name: &str) -> VmResult<usize> {
 
 // ---------------------------------------------------------------------------
 // StringBuilder natives
-// ---------------------------------------------------------------------------
-
-/// Native: `StringBuilder.<init>()V` — initialise empty buffer.
-fn native_sb_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let obj = heap.get_mut(this_ref)?;
-    obj.string_value = Some(String::new());
-    Ok(None)
-}
-
-/// Native: `StringBuilder.<init>(Ljava/lang/String;)V` — init with string.
-fn native_sb_init_string(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let init_str = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => heap.get(*r)?.string_value.clone().unwrap_or_default(),
-        _ => String::new(),
-    };
-    let obj = heap.get_mut(this_ref)?;
-    obj.string_value = Some(init_str);
-    Ok(None)
-}
-
-/// Native: `StringBuilder.append(Ljava/lang/String;)Ljava/lang/StringBuilder;`
-fn native_sb_append_string(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let append_str = match args.get(1) {
-        Some(Slot::Reference(Some(r))) => heap.get(*r)?.string_value.clone().unwrap_or_default(),
-        _ => "null".to_string(),
-    };
-    let obj = heap.get_mut(this_ref)?;
-    if let Some(ref mut buf) = obj.string_value {
-        buf.push_str(&append_str);
-    }
-    Ok(Some(Slot::Reference(Some(this_ref))))
-}
-
-/// Native: `StringBuilder.append(I)Ljava/lang/StringBuilder;`
-fn native_sb_append_int(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let val = match args.get(1) {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let obj = heap.get_mut(this_ref)?;
-    if let Some(ref mut buf) = obj.string_value {
-        buf.push_str(&val.to_string());
-    }
-    Ok(Some(Slot::Reference(Some(this_ref))))
-}
-
-/// Native: `StringBuilder.append(J)Ljava/lang/StringBuilder;`
-fn native_sb_append_long(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let val = match args.get(1) {
-        Some(Slot::Long(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Long",
-                got: "other",
-            });
-        }
-    };
-    let obj = heap.get_mut(this_ref)?;
-    if let Some(ref mut buf) = obj.string_value {
-        buf.push_str(&val.to_string());
-    }
-    Ok(Some(Slot::Reference(Some(this_ref))))
-}
-
-/// Native: `StringBuilder.append(D)Ljava/lang/StringBuilder;`
-fn native_sb_append_double(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let val = match args.get(1) {
-        Some(Slot::Double(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Double",
-                got: "other",
-            });
-        }
-    };
-    let obj = heap.get_mut(this_ref)?;
-    if let Some(ref mut buf) = obj.string_value {
-        buf.push_str(&val.to_string());
-    }
-    Ok(Some(Slot::Reference(Some(this_ref))))
-}
-
-/// Native: `StringBuilder.append(F)Ljava/lang/StringBuilder;`
-fn native_sb_append_float(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let val = match args.get(1) {
-        Some(Slot::Float(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Float",
-                got: "other",
-            });
-        }
-    };
-    let obj = heap.get_mut(this_ref)?;
-    if let Some(ref mut buf) = obj.string_value {
-        buf.push_str(&val.to_string());
-    }
-    Ok(Some(Slot::Reference(Some(this_ref))))
-}
-
-/// Native: `StringBuilder.append(Z)Ljava/lang/StringBuilder;`
-fn native_sb_append_boolean(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let val = match args.get(1) {
-        Some(Slot::Int(v)) => *v != 0,
-        _ => false,
-    };
-    let obj = heap.get_mut(this_ref)?;
-    if let Some(ref mut buf) = obj.string_value {
-        buf.push_str(if val { "true" } else { "false" });
-    }
-    Ok(Some(Slot::Reference(Some(this_ref))))
-}
-
-/// Native: `StringBuilder.append(C)Ljava/lang/StringBuilder;`
-fn native_sb_append_char(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let val = match args.get(1) {
-        Some(Slot::Int(v)) => char::from_u32((*v).cast_unsigned()).unwrap_or('\0'),
-        _ => '\0',
-    };
-    let obj = heap.get_mut(this_ref)?;
-    if let Some(ref mut buf) = obj.string_value {
-        buf.push(val);
-    }
-    Ok(Some(Slot::Reference(Some(this_ref))))
-}
-
-/// Native: `StringBuilder.toString()Ljava/lang/String;`
-fn native_sb_tostring(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let content = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
-    let r = heap.allocate_string(content);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `StringBuilder.length()I`
-fn native_sb_length(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let len = heap
-        .get(this_ref)?
-        .string_value
-        .as_ref()
-        .map_or(0, String::len);
-    Ok(Some(Slot::Int(i32::try_from(len).unwrap_or(i32::MAX))))
-}
-
-// Character natives
-// ---------------------------------------------------------------------------
-
-/// Helper: extract a `char` from a `Slot::Int` argument.
-fn slot_to_char(slot: &Slot) -> VmResult<char> {
-    match slot {
-        Slot::Int(v) => Ok(char::from_u32((*v).cast_unsigned()).unwrap_or('\0')),
-        _ => Err(VmError::TypeMismatch {
-            expected: "Int (char)",
-            got: "other",
-        }),
-    }
-}
-
-/// Native: `Character.isDigit(C)Z`
-fn native_char_is_digit(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let ch = slot_to_char(args.first().ok_or(VmError::StackUnderflow)?)?;
-    Ok(Some(Slot::Int(i32::from(ch.is_ascii_digit()))))
-}
-
-/// Native: `Character.isLetter(C)Z`
-fn native_char_is_letter(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let ch = slot_to_char(args.first().ok_or(VmError::StackUnderflow)?)?;
-    Ok(Some(Slot::Int(i32::from(ch.is_alphabetic()))))
-}
-
-/// Native: `Character.isWhitespace(C)Z`
-fn native_char_is_whitespace(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let ch = slot_to_char(args.first().ok_or(VmError::StackUnderflow)?)?;
-    Ok(Some(Slot::Int(i32::from(ch.is_whitespace()))))
-}
-
-/// Native: `Character.isUpperCase(C)Z`
-fn native_char_is_uppercase(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let ch = slot_to_char(args.first().ok_or(VmError::StackUnderflow)?)?;
-    Ok(Some(Slot::Int(i32::from(ch.is_uppercase()))))
-}
-
-/// Native: `Character.isLowerCase(C)Z`
-fn native_char_is_lowercase(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let ch = slot_to_char(args.first().ok_or(VmError::StackUnderflow)?)?;
-    Ok(Some(Slot::Int(i32::from(ch.is_lowercase()))))
-}
-
-/// Native: `Character.toUpperCase(C)C`
-fn native_char_to_uppercase(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let ch = slot_to_char(args.first().ok_or(VmError::StackUnderflow)?)?;
-    let upper = ch.to_uppercase().next().unwrap_or(ch);
-    Ok(Some(Slot::Int(upper as i32)))
-}
-
-/// Native: `Character.toLowerCase(C)C`
-fn native_char_to_lowercase(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let ch = slot_to_char(args.first().ok_or(VmError::StackUnderflow)?)?;
-    let lower = ch.to_lowercase().next().unwrap_or(ch);
-    Ok(Some(Slot::Int(lower as i32)))
-}
-
-/// Native: `Character.isLetterOrDigit(C)Z`
-fn native_char_is_letter_or_digit(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let ch = slot_to_char(args.first().ok_or(VmError::StackUnderflow)?)?;
-    Ok(Some(Slot::Int(i32::from(ch.is_alphanumeric()))))
-}
-
-/// Native: `Character.valueOf(C)Ljava/lang/Character;` — box a char.
-fn native_char_valueof(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let val = match args.first() {
-        Some(Slot::Int(v)) => *v,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int (char)",
-                got: "other",
-            });
-        }
-    };
-    let r = heap.allocate("java/lang/Character".to_string(), 1);
-    heap.get_mut(r).unwrap().fields[0] = Slot::Int(val);
-    Ok(Some(Slot::Reference(Some(r))))
-}
-
-/// Native: `Character.charValue()C` — unbox Character to char.
-fn native_char_charvalue(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let val = heap.get(this_ref)?.fields[0];
-    Ok(Some(val))
-}
-
-// ---------------------------------------------------------------------------
-// ArrayList natives
-// ---------------------------------------------------------------------------
-
-/// Native: `ArrayList.<init>()V` — initializes with size=0.
-fn native_arraylist_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    heap.get_mut(this_ref)?.fields[0] = Slot::Int(0);
-    Ok(None)
-}
-
-/// Native: `ArrayList.add(Object)Z` — appends element, returns true.
-fn native_arraylist_add(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let element = args.get(1).copied().unwrap_or(Slot::Reference(None));
-    let obj = heap.get_mut(this_ref)?;
-    match obj.fields.first_mut() {
-        Some(Slot::Int(sz)) => *sz += 1,
-        _ => return Err(VmError::NullPointerException), // shouldn't happen; init sets fields[0]=Int(0)
-    }
-    obj.fields.push(element);
-    Ok(Some(Slot::Int(1))) // boolean true
-}
-
-/// Native: `ArrayList.get(I)Object` — returns element at index.
-#[allow(clippy::cast_sign_loss)]
-fn native_arraylist_get(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let idx = match args.get(1) {
-        Some(Slot::Int(i)) => *i as usize,
-        _ => {
-            return Err(VmError::TypeMismatch {
-                expected: "Int",
-                got: "other",
-            });
-        }
-    };
-    let obj = heap.get(this_ref)?;
-    obj.fields.get(idx + 1).map_or_else(
-        || {
-            Err(VmError::JavaException {
-                class_name: "java/lang/ArrayIndexOutOfBoundsException".to_string(),
-            })
-        },
-        |slot| Ok(Some(*slot)),
-    )
-}
-
-/// Native: `ArrayList.size()I`
-fn native_arraylist_size(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let obj = heap.get(this_ref)?;
-    match obj.fields.first() {
-        Some(Slot::Int(sz)) => Ok(Some(Slot::Int(*sz))),
-        _ => Ok(Some(Slot::Int(0))),
-    }
-}
-
-/// Native: `ArrayList.iterator()Iterator` — creates an `ArrayListIterator`.
-fn native_arraylist_iterator(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let iter_ref = heap.allocate("duke/util/ArrayListIterator".to_string(), 2);
-    {
-        let iter_obj = heap.get_mut(iter_ref)?;
-        iter_obj.fields[0] = Slot::Reference(Some(this_ref));
-        iter_obj.fields[1] = Slot::Int(0);
-    }
-    Ok(Some(Slot::Reference(Some(iter_ref))))
-}
-
-/// Native: `ArrayList.sort(Comparator)V` — sorts in-place using insertion sort,
-/// calling `compareTo` on each element pair via the interpreter callback.
-///
-/// Only null Comparator (natural ordering via `compareTo`) is supported.
-fn array_list_sort(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    output: &mut dyn Write,
-    _control: &mut NativeControl,
-    invoke: &mut InvokeFn<'_>,
-) -> VmResult<Option<Slot>> {
-    // args[0] = ArrayList ref, args[1] = Comparator (null = natural ordering)
-    let list_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-
-    // Fix 1: use the correct error variant for unsupported non-null Comparator.
-    if !matches!(args.get(1), Some(Slot::Reference(None)) | None) {
-        return Err(VmError::Unimplemented {
-            mnemonic: "ArrayList.sort(non-null Comparator)",
-        });
-    }
-
-    // Fix 2: guard against a negative size stored in fields[0].
-    let size = match heap.get(list_ref)?.fields.first() {
-        Some(Slot::Int(n)) if *n >= 0 => usize::try_from(*n).unwrap_or(0),
-        Some(Slot::Int(n)) => return Err(VmError::NegativeArraySize { size: *n }),
-        _ => return Ok(None),
-    };
-
-    if size <= 1 {
-        return Ok(None);
-    }
-
-    // Collect element refs (fields[1..=size]).
-    let mut elems: Vec<u64> = (1..=size)
-        .filter_map(|i| match heap.get(list_ref).ok()?.fields.get(i) {
-            Some(Slot::Reference(Some(r))) => Some(*r),
-            _ => None,
-        })
-        .collect();
-
-    // Fix 3: malformed list → InvalidRef, not silent Ok(None).
-    if elems.len() != size {
-        return Err(VmError::InvalidRef { address: list_ref });
-    }
-
-    // Insertion sort — O(n²), correct, easy to verify.
-    for i in 1..elems.len() {
-        let mut key = elems[i];
-        let mut j = i;
-        while j > 0 {
-            let receiver = elems[j - 1];
-            let class_name = heap.get(receiver)?.class_name.clone();
-            let cmp = invoke(
-                heap,
-                output,
-                &class_name,
-                COMPARE_TO_METHOD,
-                COMPARE_TO_OBJECT_DESC,
-                vec![Slot::Reference(Some(receiver)), Slot::Reference(Some(key))],
-            )?;
-
-            // Patch stale young-gen refs if a minor GC fired during the callback.
-            // Guarded by `has_pending_forwards` so the common (no-GC) path pays
-            // only one bool check instead of O(n) HashMap probes.
-            if heap.has_pending_forwards() {
-                for elem in &mut elems {
-                    let mut slot = Slot::Reference(Some(*elem));
-                    heap.apply_forward(&mut slot);
-                    if let Slot::Reference(Some(r)) = slot {
-                        *elem = r;
-                    }
-                }
-                // Re-read key after forwarding patch (it lives outside elems
-                // during the innermost loop iteration).
-                let mut key_slot = Slot::Reference(Some(key));
-                heap.apply_forward(&mut key_slot);
-                if let Slot::Reference(Some(r)) = key_slot {
-                    key = r;
-                }
-            }
-
-            // Fix 4: explicit error on non-Int compareTo return.
-            match cmp {
-                Some(Slot::Int(n)) if n <= 0 => break,
-                Some(Slot::Int(_)) => {} // n > 0, keep shifting
-                _ => {
-                    return Err(VmError::TypeMismatch {
-                        expected: "Int",
-                        got: "other",
-                    });
-                }
-            }
-            elems[j] = elems[j - 1];
-            j -= 1;
-        }
-        elems[j] = key;
-    }
-
-    // Write sorted elements back using the write barrier.
-    for (i, &r) in elems.iter().enumerate() {
-        heap.write_field(list_ref, i + 1, Slot::Reference(Some(r)))?;
-    }
-
-    Ok(None)
-}
-
-// ---------------------------------------------------------------------------
-// Collections natives
-// ---------------------------------------------------------------------------
-
-/// Native: `Collections.sort(List)V` — delegates to the list's sort(null) method.
-///
-/// `Collections.sort(list)` is compiled by javac as
-/// `invokestatic java/util/Collections.sort:(Ljava/util/List;)V`.
-/// We forward to the runtime class's `sort(Comparator=null)`, which for an
-/// `ArrayList` performs the insertion-sort-with-compareTo callback.
-fn native_collections_sort(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    output: &mut dyn Write,
-    _control: &mut NativeControl,
-    invoke: &mut InvokeFn<'_>,
-) -> VmResult<Option<Slot>> {
-    // args[0] = List ref
-    let list_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    // Dispatch on the actual runtime class so any List implementation works.
-    let class_name = heap.get(list_ref)?.class_name.clone();
-    invoke(
-        heap,
-        output,
-        &class_name,
-        "sort",
-        SORT_COMPARATOR_DESC,
-        vec![Slot::Reference(Some(list_ref)), Slot::Reference(None)],
-    )?;
-    Ok(None)
-}
-
-// ---------------------------------------------------------------------------
-// ArrayListIterator natives
-// ---------------------------------------------------------------------------
-
-/// Native: `ArrayListIterator.<init>` — no-op; fields set directly by `native_arraylist_iterator`.
-#[allow(clippy::unnecessary_wraps)]
-fn native_arraylist_iter_init(
-    _args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    Ok(None)
-}
-
-/// Native: `ArrayListIterator.hasNext()Z`
-fn native_arraylist_iter_hasnext(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let iter_obj = heap.get(this_ref)?;
-    let list_ref = match iter_obj.fields.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Ok(Some(Slot::Int(0))),
-    };
-    let cursor = match iter_obj.fields.get(1) {
-        Some(Slot::Int(i)) => *i,
-        _ => 0,
-    };
-    let list_size = match heap.get(list_ref)?.fields.first() {
-        Some(Slot::Int(sz)) => *sz,
-        _ => 0,
-    };
-    Ok(Some(Slot::Int(i32::from(cursor < list_size))))
-}
-
-/// Native: `ArrayListIterator.next()Object` — returns element at cursor, advances cursor.
-#[allow(clippy::cast_sign_loss)]
-fn native_arraylist_iter_next(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let (list_ref, cursor) = {
-        let iter_obj = heap.get(this_ref)?;
-        let lr = match iter_obj.fields.first() {
-            Some(Slot::Reference(Some(r))) => *r,
-            _ => return Err(VmError::NullPointerException),
-        };
-        let c = match iter_obj.fields.get(1) {
-            Some(Slot::Int(i)) => *i,
-            _ => 0,
-        };
-        (lr, c)
-    };
-    let element = {
-        let list_obj = heap.get(list_ref)?;
-        match list_obj.fields.get(cursor as usize + 1) {
-            Some(slot) => *slot,
-            None => {
-                return Err(VmError::JavaException {
-                    class_name: "java/util/NoSuchElementException".to_string(),
-                });
-            }
-        }
-    };
-    heap.get_mut(this_ref)?.fields[1] = Slot::Int(cursor + 1);
-    Ok(Some(element))
-}
-
-// ---- Double.isNaN ----
-
-/// Native: `Double.isNaN(D)Z` — returns 1 if value is NaN.
-#[allow(clippy::unnecessary_wraps)]
-fn native_double_isnan(
-    args: &[Slot],
-    _heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    match args.first() {
-        Some(Slot::Double(v)) => Ok(Some(Slot::Int(i32::from(v.is_nan())))),
-        _ => Ok(Some(Slot::Int(0))),
-    }
-}
-
-/// Native: `Double.compareTo(Object)` — compares two boxed Doubles.
-fn native_double_compareto(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let double_val = |s: &Slot| -> VmResult<f64> {
-        match s {
-            Slot::Reference(Some(r)) => match heap.get(*r)?.fields.first() {
-                Some(Slot::Double(n)) => Ok(*n),
-                _ => Err(VmError::InvalidRef { address: *r }),
-            },
-            _ => Err(VmError::NullPointerException),
-        }
-    };
-    let a = match args.first() {
-        Some(s) => double_val(s)?,
-        None => return Err(VmError::NullPointerException),
-    };
-    let b = match args.get(1) {
-        Some(s) => double_val(s)?,
-        None => return Err(VmError::NullPointerException),
-    };
-    // Use total_cmp: implements Java's total order where NaN > +∞ > … > -∞.
-    Ok(Some(Slot::Int(ordering_to_int(a.total_cmp(&b)))))
-}
-
-// ---- Arrays natives ----
-
-/// Native: `Arrays.fill(int[], int)` — fills all elements with val.
-fn native_arrays_fill_int(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let arr_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let val = match args.get(1) {
-        Some(Slot::Int(v)) => Slot::Int(*v),
-        _ => Slot::Int(0),
-    };
-    let obj = heap.get_mut(arr_ref)?;
-    for slot in &mut obj.fields {
-        *slot = val;
-    }
-    Ok(None)
-}
-
-/// Native: `Arrays.fill(Object[], Object)` — fills all elements with val.
-fn native_arrays_fill_object(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let arr_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let val = args.get(1).copied().unwrap_or(Slot::Reference(None));
-    let obj = heap.get_mut(arr_ref)?;
-    for slot in &mut obj.fields {
-        *slot = val;
-    }
-    Ok(None)
-}
-
-/// Native: `Arrays.copyOf(int[], int)` — copies to new int[] of given length.
-fn native_arrays_copyof_int(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let src_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let new_len = match args.get(1) {
-        Some(Slot::Int(n)) if *n >= 0 => usize::try_from(*n).unwrap_or(0),
-        Some(Slot::Int(n)) => return Err(VmError::NegativeArraySize { size: *n }),
-        _ => 0,
-    };
-    let src_fields = heap.get(src_ref)?.fields.clone();
-    let dst_ref = heap.allocate("[I".to_string(), new_len);
-    let dst = heap.get_mut(dst_ref)?;
-    for i in 0..new_len {
-        dst.fields[i] = src_fields.get(i).copied().unwrap_or(Slot::Int(0));
-    }
-    Ok(Some(Slot::Reference(Some(dst_ref))))
-}
-
-/// Native: `Arrays.copyOf(Object[], int)` — copies to new Object[] of given length.
-fn native_arrays_copyof_object(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let src_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let new_len = match args.get(1) {
-        Some(Slot::Int(n)) if *n >= 0 => usize::try_from(*n).unwrap_or(0),
-        Some(Slot::Int(n)) => return Err(VmError::NegativeArraySize { size: *n }),
-        _ => 0,
-    };
-    let src_fields = heap.get(src_ref)?.fields.clone();
-    let dst_ref = heap.allocate("[Ljava/lang/Object;".to_string(), new_len);
-    let dst = heap.get_mut(dst_ref)?;
-    for i in 0..new_len {
-        dst.fields[i] = src_fields.get(i).copied().unwrap_or(Slot::Reference(None));
-    }
-    Ok(Some(Slot::Reference(Some(dst_ref))))
-}
-
-/// Native: `Arrays.sort(int[])` — sorts fields in place.
-fn native_arrays_sort_int(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let arr_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let obj = heap.get_mut(arr_ref)?;
-    obj.fields.sort_by(|a, b| match (a, b) {
-        (Slot::Int(x), Slot::Int(y)) => x.cmp(y),
-        _ => std::cmp::Ordering::Equal,
-    });
-    Ok(None)
-}
-
-// ---------------------------------------------------------------------------
-// HashMap natives
-// ---------------------------------------------------------------------------
-
-/// Semantic equality for `HashMap` keys: compares by `string_value` for heap strings,
-/// or by the first field (e.g. intValue) for boxed numerics, or by reference identity.
-fn slots_equal(a: &Slot, b: &Slot, heap: &duke_gc::Heap) -> bool {
-    match (a, b) {
-        (Slot::Reference(None), Slot::Reference(None)) => true,
-        (Slot::Reference(Some(ra)), Slot::Reference(Some(rb))) => {
-            if ra == rb {
-                return true;
-            }
-            let Ok(oa) = heap.get(*ra) else { return false };
-            let Ok(ob) = heap.get(*rb) else { return false };
-            if oa.class_name == "java/lang/String" || ob.class_name == "java/lang/String" {
-                return oa.string_value == ob.string_value;
-            }
-            if oa.class_name == ob.class_name {
-                oa.fields.first() == ob.fields.first()
-            } else {
-                false
-            }
-        }
-        _ => false,
-    }
-}
-
-/// Native: `HashMap.<init>()V` — initialises size counter at fields\[0\] to 0.
-fn find_hashmap_entry_index(fields: &[Slot], key: &Slot, heap: &duke_gc::Heap) -> Option<usize> {
-    (1..fields.len())
-        .step_by(2)
-        .find(|&i| i + 1 < fields.len() && slots_equal(&fields[i], key, heap))
-}
-
-fn native_hashmap_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let obj = heap.get_mut(this_ref)?;
-    if obj.fields.is_empty() {
-        obj.fields.push(Slot::Int(0));
-    } else {
-        obj.fields[0] = Slot::Int(0);
-    }
-    Ok(None)
-}
-
-/// Native: `HashMap.put(Object, Object)Object` — inserts or updates a key-value pair.
-/// Returns the old value if the key was already present, or null if it is new.
-fn native_hashmap_put(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let key = args.get(1).copied().unwrap_or(Slot::Reference(None));
-    let val = args.get(2).copied().unwrap_or(Slot::Reference(None));
-    // Clone fields to release the immutable borrow before mutating.
-    let fields = heap.get(this_ref)?.fields.clone();
-
-    if let Some(i) = find_hashmap_entry_index(&fields, &key, heap) {
-        let old = fields[i + 1];
-        heap.get_mut(this_ref)?.fields[i + 1] = val;
-        return Ok(Some(old));
-    }
-
-    // New key — append pair and bump size.
-    let obj = heap.get_mut(this_ref)?;
-    match obj.fields.first_mut() {
-        Some(Slot::Int(sz)) => *sz += 1,
-        _ => return Err(VmError::NullPointerException),
-    }
-    obj.fields.push(key);
-    obj.fields.push(val);
-    Ok(Some(Slot::Reference(None)))
-}
-
-/// Native: `HashMap.get(Object)Object` — returns value for key, or null if absent.
-fn native_hashmap_get(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let key = args.get(1).copied().unwrap_or(Slot::Reference(None));
-    let fields = heap.get(this_ref)?.fields.clone();
-
-    Ok(Some(
-        find_hashmap_entry_index(&fields, &key, heap)
-            .map_or(Slot::Reference(None), |i| fields[i + 1]),
-    ))
-}
-
-/// Native: `HashMap.containsKey(Object)Z` — returns 1 if key present, 0 otherwise.
-fn native_hashmap_contains_key(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let key = args.get(1).copied().unwrap_or(Slot::Reference(None));
-    let fields = heap.get(this_ref)?.fields.clone();
-
-    if find_hashmap_entry_index(&fields, &key, heap).is_some() {
-        Ok(Some(Slot::Int(1)))
-    } else {
-        Ok(Some(Slot::Int(0)))
-    }
-}
-
-/// Native: `HashMap.size()I` — returns entry count from fields\[0\].
-fn native_hashmap_size(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(n)) => Ok(Some(Slot::Int(*n))),
-        _ => Ok(Some(Slot::Int(0))),
-    }
-}
-
-/// Native: `HashMap.remove(Object)Object` — removes a key-value pair, returns old value or null.
-/// Uses swap-remove (swaps target pair with last pair) for O(1) deletion.
-fn native_hashmap_remove(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let key = args.get(1).copied().unwrap_or(Slot::Reference(None));
-    let fields = heap.get(this_ref)?.fields.clone();
-
-    if let Some(i) = find_hashmap_entry_index(&fields, &key, heap) {
-        let old_val = fields[i + 1];
-        let obj = heap.get_mut(this_ref)?;
-        let last_val_idx = obj.fields.len() - 1;
-        let last_key_idx = obj.fields.len() - 2;
-        obj.fields.swap(i + 1, last_val_idx);
-        obj.fields.swap(i, last_key_idx);
-        obj.fields.truncate(obj.fields.len() - 2);
-        match obj.fields.first_mut() {
-            Some(Slot::Int(sz)) => *sz -= 1,
-            _ => return Err(VmError::NullPointerException),
-        }
-        Ok(Some(old_val))
-    } else {
-        Ok(Some(Slot::Reference(None)))
-    }
-}
-
-/// Native: `HashMap.isEmpty()Z` — returns 1 if size == 0, else 0.
-fn native_hashmap_is_empty(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(0)) => Ok(Some(Slot::Int(1))),
-        Some(Slot::Int(_)) => Ok(Some(Slot::Int(0))),
-        _ => Ok(Some(Slot::Int(1))),
-    }
-}
-
-/// Native: `HashMap.getOrDefault(Object, Object)Object` — returns value for key, or default if absent.
-fn native_hashmap_get_or_default(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let key = args.get(1).copied().unwrap_or(Slot::Reference(None));
-    let default = args.get(2).copied().unwrap_or(Slot::Reference(None));
-    let fields = heap.get(this_ref)?.fields.clone();
-
-    Ok(Some(
-        find_hashmap_entry_index(&fields, &key, heap).map_or(default, |i| fields[i + 1]),
-    ))
-}
-
-// ---------------------------------------------------------------------------
-// HashSet natives
-// ---------------------------------------------------------------------------
-
-/// Native: `HashSet.<init>()V` — initialises size counter at fields\[0\] to 0.
-fn find_hashset_entry_index(
-    fields: &[Slot],
-    element: &Slot,
-    heap: &duke_gc::Heap,
-) -> Option<usize> {
-    fields
-        .iter()
-        .skip(1)
-        .position(|field| slots_equal(field, element, heap))
-        .map(|idx| idx + 1)
-}
-
-fn native_hashset_init(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let obj = heap.get_mut(this_ref)?;
-    if obj.fields.is_empty() {
-        obj.fields.push(Slot::Int(0));
-    } else {
-        obj.fields[0] = Slot::Int(0);
-    }
-    Ok(None)
-}
-
-/// Native: `HashSet.add(Object)Z` — adds element if not already present.
-/// Returns 1 if added, 0 if element was already in the set.
-fn native_hashset_add(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let element = args.get(1).copied().unwrap_or(Slot::Reference(None));
-    let fields = heap.get(this_ref)?.fields.clone();
-    // fields[0] = size, fields[1..] = elements
-    if find_hashset_entry_index(&fields, &element, heap).is_some() {
-        return Ok(Some(Slot::Int(0))); // duplicate
-    }
-
-    let obj = heap.get_mut(this_ref)?;
-    match obj.fields.first_mut() {
-        Some(Slot::Int(sz)) => *sz += 1,
-        _ => return Err(VmError::NullPointerException),
-    }
-    obj.fields.push(element);
-    Ok(Some(Slot::Int(1)))
-}
-
-/// Native: `HashSet.contains(Object)Z` — returns 1 if element is present, 0 otherwise.
-fn native_hashset_contains(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let element = args.get(1).copied().unwrap_or(Slot::Reference(None));
-    let fields = heap.get(this_ref)?.fields.clone();
-
-    if find_hashset_entry_index(&fields, &element, heap).is_some() {
-        Ok(Some(Slot::Int(1)))
-    } else {
-        Ok(Some(Slot::Int(0)))
-    }
-}
-
-/// Native: `HashSet.remove(Object)Z` — removes element if present, returns 1 if removed, 0 if absent.
-/// Uses swap-remove (swaps target with last element) for O(1) deletion.
-fn native_hashset_remove(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    let element = args.get(1).copied().unwrap_or(Slot::Reference(None));
-    let fields = heap.get(this_ref)?.fields.clone();
-
-    if let Some(i) = find_hashset_entry_index(&fields, &element, heap) {
-        let obj = heap.get_mut(this_ref)?;
-        let last_idx = obj.fields.len() - 1;
-        obj.fields.swap(i, last_idx);
-        obj.fields.truncate(obj.fields.len() - 1);
-        match obj.fields.first_mut() {
-            Some(Slot::Int(sz)) => *sz -= 1,
-            _ => return Err(VmError::NullPointerException),
-        }
-        Ok(Some(Slot::Int(1)))
-    } else {
-        Ok(Some(Slot::Int(0)))
-    }
-}
-
-/// Native: `HashSet.size()I` — returns element count from fields\[0\].
-fn native_hashset_size(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(n)) => Ok(Some(Slot::Int(*n))),
-        _ => Ok(Some(Slot::Int(0))),
-    }
-}
-
-/// Native: `HashSet.isEmpty()Z` — returns 1 if size == 0, else 0.
-fn native_hashset_is_empty(
-    args: &[Slot],
-    heap: &mut duke_gc::Heap,
-    _out: &mut dyn Write,
-    _control: &mut NativeControl,
-) -> VmResult<Option<Slot>> {
-    let this_ref = match args.first() {
-        Some(Slot::Reference(Some(r))) => *r,
-        _ => return Err(VmError::NullPointerException),
-    };
-    match heap.get(this_ref)?.fields.first() {
-        Some(Slot::Int(0)) => Ok(Some(Slot::Int(1))),
-        Some(Slot::Int(_)) => Ok(Some(Slot::Int(0))),
-        _ => Ok(Some(Slot::Int(1))),
-    }
-}
-
-// ---------------------------------------------------------------------------
-// GC root gathering
-// ---------------------------------------------------------------------------
-
-/// Collect all live Slot values from the interpreter's current execution state.
-/// The GC uses these as the root set for reachability analysis.
-fn gather_roots(
-    frame: &duke_runtime::Frame,
-    call_stack: &[CallFrame],
-    registry: &ClassRegistry,
-) -> Vec<duke_runtime::Slot> {
-    let mut roots = Vec::new();
-    roots.extend(frame.slots());
-    for cf in call_stack {
-        roots.extend(cf.frame.slots());
-    }
-    for ctx in registry.all_classes() {
-        roots.extend(ctx.static_fields.iter().copied());
-    }
-    roots
-}
-
-/// Apply GC forwarding pointers to all live interpreter slots after a minor
-/// collection. Must be called immediately after `heap.collect()` returns so
-/// that stale young-gen references are updated to their new locations.
-fn patch_forwarded_slots(
-    frame: &mut duke_runtime::Frame,
-    call_stack: &mut [CallFrame],
-    registry: &mut ClassRegistry,
-    heap: &duke_gc::Heap,
-) {
-    for slot in frame.slots_mut() {
-        heap.apply_forward(slot);
-    }
-    for cf in call_stack.iter_mut() {
-        for slot in cf.frame.slots_mut() {
-            heap.apply_forward(slot);
-        }
-    }
-    for ctx in registry.all_classes_mut() {
-        for slot in &mut ctx.static_fields {
-            heap.apply_forward(slot);
-        }
-    }
-}
 
 // ---------------------------------------------------------------------------
 // Tests
@@ -11826,6 +7613,7 @@ fn patch_forwarded_slots(
 
 #[cfg(test)]
 mod tests {
+    use super::natives::*;
     use super::*;
 
     macro_rules! wrap_simple_native_for_tests {
@@ -11836,7 +7624,7 @@ mod tests {
                     heap: &mut duke_gc::Heap,
                     out: &mut dyn std::io::Write,
                 ) -> VmResult<Option<Slot>> {
-                    super::$name(args, heap, out, &mut NativeControl::default())
+                    super::natives::$name(args, heap, out, &mut NativeControl::default())
                 }
             )*
         };
@@ -11894,7 +7682,7 @@ mod tests {
         out: &mut dyn std::io::Write,
         invoke: &mut InvokeFn<'_>,
     ) -> VmResult<Option<Slot>> {
-        super::array_list_sort(args, heap, out, &mut NativeControl::default(), invoke)
+        super::natives::array_list_sort(args, heap, out, &mut NativeControl::default(), invoke)
     }
 
     // ---- Unit tests: hand-crafted instruction streams ----
